% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Get abundances}
\usage{
abundance(data)
}
\arguments{
\item{data}{d}
}
\value{
two-dimensional \code{matrix} of mode \code{numeric};
}
\description{
S4 generic function \code{abundance()}.
}
\details{

}
\examples{

}

