% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\name{additive.to.diversity}
\alias{additive.to.diversity}
\title{additive.to.diversity}
\usage{
additive.to.diversity(additive)
}
\arguments{
\item{additive}{Dataframe comprising 'additive diversity' values, whereby 
each row and column corresponds to a particular subcommunity and q value, 
respectively; q values should be input in column names as character 
strings, e.g. "q1".}
}
\value{
Dataframe comprising diversity values, whereby each row and column 
corresponds to a particular subcommunity and q value, respectively
}
\description{
Converts 'additive diversity' values (for any q) to diversities.
}

