% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{alpha}
\alias{alpha}
\title{Low level diversity components: alpha}
\usage{
alpha(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns
}
\description{
Calculates the low-level diversity component necessary for calculating alpha
diversity.
}
\details{
Values generated from \code{alpha()} may be input into \code{subdiv()} and
\code{superdiv} to calcualte raw subcommunity/supercommunity alpha diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate alpha component
a <- alpha(super)

}

