% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.collection.R
\name{as.collection}
\alias{as.collection}
\alias{is.collection}
\title{Coerse to a Collection}
\usage{
as.collection(data, similarity = NA, zmatrix = NA, lookup = NA)

is.collection(x)
}
\arguments{
\item{data}{Object of class \code{matrix}}

\item{similarity}{Object of class \code{character}}

\item{zmatrix}{Object of class \code{matrix}}

\item{lookup}{Object of class \code{data.frame}}

\item{x}{any R object}
}
\value{
\code{as.collection()} returns an object of class \code{collection}; 
an S4 object containing two slots, pmatrix and zmatrix. \cr\cr
\code{is.collection()} returns TRUE if its argument is a collection, 
FALSE otherwise.
}
\description{
Functions to check if an object is a \code{collection} or coerce an object 
into a \code{collection}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class collection
data <- as.collection(population)

print(data)
print(data@zmatrix)

}

