% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rdphylo.R
\name{as.RDphylo}
\alias{as.RDphylo}
\alias{is.RDphylo}
\title{Phylogenetic data}
\usage{
as.RDphylo(tree, pds.abundance = matrix(rep(1/length(tree$tip.label),
  length(tree$tip.label))))

is.RDphylo(x)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{pds.abundance}{\code{vector} of length \eqn{S}; proportional
abundance of present-day species}

\item{x}{any R object}
}
\value{
S4 generic of class \linkS4class{RDphylo} containing historic 
species: names, ancestral and descendant nodes, descendant present day 
species, and proportional abundance.
}
\description{
Functions to check if an object is of class \code{RDphylo} or coerce an 
object into class \code{RDphylo}.
}

