% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.supercommunity.R
\name{as.supercommunity}
\alias{as.supercommunity}
\alias{is.supercommunity}
\title{Coerse to a Supercommunity}
\usage{
as.supercommunity(data, similarity = NA, zmatrix = NA, lookup = NA)

is.supercommunity(x)
}
\arguments{
\item{data}{Object of class \code{matrix}}

\item{similarity}{Object of class \code{character}}

\item{zmatrix}{Object of class \code{matrix}}

\item{lookup}{Object of class \code{data.frame}}

\item{x}{any R object}
}
\value{
\code{as.supercommunity()} returns an object of class \code{supercommunity}; 
an S4 object containing two slots, pmatrix and zmatrix. \cr\cr
\code{is.supercommunity()} returns TRUE if its argument is a supercommunity, 
FALSE otherwise.
}
\description{
Functions to check if an object is a \code{supercommunity} or coerce an object 
into a \code{supercommunity}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class supercommunity
data <- as.supercommunity(population)

print(data)
print(data@zmatrix)

}

