% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{beta}
\alias{beta}
\title{Low level diversity components: gamma}
\usage{
beta(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns
}
\description{
Calculates the low-level diversity component necessary for calculating beta
diversity.
}
\details{
Values generated from \code{beta()} may be input into \code{subdiv()} and
\code{superdiv} to calcualte raw subcommunity/supercommunity beta diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate beta component
a <- beta(super)

}

