% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{betabar}
\alias{betabar}
\title{Low level diversity components}
\usage{
betabar(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns
}
\description{
Calculates the low-level diversity component necessary for calculating 
normalised beta diversity.
}
\details{
Values generated from \code{betabar()} may be input into \code{subdiv()} and
\code{superdiv} to calcualte normalised subcommunity/supercommunity beta 
diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate normalised beta component
a <- betabar(super)

}

