% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{byte}
\alias{byte}
\title{A simple helper function to return the size of one byte}
\usage{
byte()
}
\value{
The size of one byte (8)
}
\description{
Used to increase readabilaty
}
\seealso{
\link{fillUpToByte}
}
