% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rdphylo.R
\name{calc.Lj}
\alias{calc.Lj}
\title{Total evolutionary change}
\usage{
calc.Lj(tree, node)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{node}{integer corresponding to the node of interest}
}
\value{
object of class \code{numeric}
}
\description{
Calculate the total length of evolutionary change of species \emph{j}; may 
be an internal or external node corresponding to present-day and historic 
species, respectively.
}

