% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainsaw.R
\name{chainsaw}
\alias{chainsaw}
\title{Cut phylogeny}
\usage{
chainsaw(partition, ps)
}
\arguments{
\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with
rows as types, columns as subcommunities, and elements containing relative
abundances of types in subcommunities. In the case of phylogenetic
metacommunities, these are the relative abundances of terminal taxa.}

\item{ps}{\code{phy_struct()} output.}
}
\value{
Returns an object of class \code{metacommunity}.
}
\description{
Cut phylogeny
}
