% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_partition.R
\name{check_partition}
\alias{check_partition}
\title{Check partition}
\usage{
check_partition(partition)
}
\arguments{
\item{partition}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains relative abundance of types}
}
\value{
Returns a two-dimensions \code{matrix} of mode \code{numeric}. If 
the partition matrix was valid, this should be identical to that which was
input as an argument.
}
\description{
\code{check_partition()} is used to validate partition matrices.
}
\examples{
population <- data.frame(a = 1:2, b = 3:4)
population <- population / sum(population)
row.names(population) <- NULL
new.population <- check_partition(population)
new.population

}

