% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phypartition.R
\name{check_phypartition}
\alias{check_phypartition}
\title{check_phypartition}
\usage{
check_phypartition(tip_labels, partition)
}
\arguments{
\item{tip_labels}{vector containing elements of class \code{character}.}

\item{partition}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types (species), columns as subcommunities, and each
element containing the relative abundance of types in each subcommunity
relative to the metacommunity as a whole. In the phylogenetic case, this
corresponds to the proportional abundance of historical species, which is
calculated from the proportional abundance of terminal taxa}
}
\value{
Returns a two-dimensions \code{matrix} of mode \code{numeric}. If
the partition matrix was valid, this should be identical to that which was
input as an argument.
}
\description{
\code{check_phypartition()} is used to validate partition matrices for use
with phylogenies.
}
