% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-collection.R
\docType{class}
\name{collection-class}
\alias{collection}
\alias{collection-class}
\alias{show,collection-method}
\title{Class to contain proportional abundance and similarity matrices}
\usage{
\S4method{show}{collection}(object)
}
\arguments{
\item{object}{object of class \linkS4class{collection}}
}
\description{
Container for proportional abundance and similarity matrices.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Prints pmatrix
}}
\section{Fields}{

\describe{
\item{\code{.Data}}{object of class \code{matrix}; proportional abundance 
of \emph{types}}

\item{\code{zmatrix}}{object of class \code{matrix}; pairwise similarity 
between \emph{types}}
}}

