% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commdiv.R
\docType{methods}
\name{commdiv}
\alias{commdiv}
\alias{commdiv,ANY,ANY,matrix-method}
\alias{commdiv,supercommunity,ANY,ANY-method}
\title{Calculate community diversity}
\usage{
commdiv(data, qs, levels)

\S4method{commdiv}{ANY,ANY,matrix}(data, qs, levels)

\S4method{commdiv}{supercommunity,ANY,ANY}(data, qs, levels)
}
\arguments{
\item{data}{two-dimensional \code{matrix} of mode \code{numeric}; diversity 
components.}

\item{qs}{\code{vector} of mode \code{numeric}; parameter of conservatism.}
}
\value{
Returns a two-dimensional \code{matrix} of mode \code{numeric}.


}
\description{
Generic function for calculating community diversity.
}
\details{
\code{data} may be input as three different classes:
\itemize{
\item{\code{powermean} calculates comunity alpha, alphabar, rho, rhobar, 
or gamma diversity by taking the powermean of the subcommunity diversity components}
\item{\code{relativeentropy} calculates community beta or betabar 
diversity by taking the relative entropy of the subcommunity diversity components}
\item{\code{supercommunity} caculates all subcommunity measures of diversity}
}

as ds
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate subcommunity gamma diversity (takes the power mean)
g <- gamma(super)
subdiv(g, 0:2)

# Calculate subcommunity beta diversity (takes the relative entropy)
b <- beta(super)
subdiv(b, 0:2)

# Calculate all measures of subcommunity diversity
subdiv(super, 0:2)

}

