% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.axis.label.R
\name{create.axis.label}
\alias{create.axis.label}
\title{Create axis label}
\usage{
create.axis.label(results, symbol = F)
}
\arguments{
\item{results}{object of class \code{data.frame} or \code{data_frame}, output from \code{subdiv()} or \code{metadiv}}

\item{symbol}{(optional) by default, output is e.g. "Metacommunity x"; if symbol is set to \code{TRUE}, output will be given as "x}
}
\description{
Create axis label
}

