% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smatrix.R
\name{descendant_tips}
\alias{descendant_tips}
\title{descendant_tips}
\usage{
descendant_tips(tree, node)
}
\arguments{
\item{tree}{object of class \code{phylo}.}

\item{node}{object of class \code{numeric}.}
}
\description{
descendant_tips
}
