% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Generate distance object}
\usage{
distance(distance, divID, values, taxID, taxMask)
}
\arguments{
\item{distance}{lksdjf}

\item{divID}{lskdj}

\item{values}{lskdj}

\item{taxID}{lskdj}

\item{taxMask}{lskdj}
}
\value{
distance() returns an object of class \code{distance}
}
\description{
distance
}
