% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-diversity.R, R/diversity.R
\docType{class}
\name{diversity-class}
\alias{diversity-class}
\alias{show,diversity-method}
\alias{plot,diversity-method}
\title{diversity-class}
\usage{
\S4method{show}{diversity}(object)

\S4method{plot}{diversity}(x)
}
\arguments{
\item{object}{object of class \code{diversity}}

\item{x}{object of class \code{diversity}}
}
\description{
Container for class \code{diversity}.
}
