% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\docType{methods}
\name{diversity}
\alias{diversity}
\alias{diversity,data.frame-method}
\alias{diversity,list-method}
\title{Generate diversity}
\usage{
diversity(res)

\S4method{diversity}{data.frame}(res)

\S4method{diversity}{list}(res)
}
\arguments{
\item{res}{object of class \code{diversity}}
}
\value{
\code{diversity()} returns an object of class \code{diversity}.
}
\description{
Functions to check if an object is a \code{diversity}
}
\examples{
pop <- data.frame(a = c(1,3), b = c(1,1))
row.names(pop) <- paste0("sp", 1:2)
pop <- pop/sum(pop)
meta <- metacommunity(pop)
qs <- c(seq(0,1,.1),2:10, seq(20,100,10),Inf)

b <- raw_beta(meta)
sc <- subdiv(b, qs)
mc <- metadiv(b, qs)
class(sc)
class(mc)

res <- rbind.data.frame(sc, mc)
res <- diversity(res)
class(res)

res <- list(sc, mc)
res <- diversity(res)
class(res)

}
