% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{allMeasures}
\alias{diversity}
\alias{selectMeasures}
\alias{subMeasures}
\alias{superMeasures}
\title{Calculate Diversity}
\usage{
diversity(measure, data, qs)

selectMeasures(measure, data, qs)

subMeasures(data, qs)

superMeasures(data, qs)

allMeasures(data, qs)
}
\arguments{
\item{measure}{function or list of functions; see Details, below}

\item{data}{object of class \code{supercommunity}; containing pmatrix and 
zmatrix}

\item{qs}{\code{vector} of mode \code{numeric}; giving \emph{q} values}
}
\value{
A list is returned of length equal to the number of diversity 
measures calculated. Each list item contains an array of diversities. 
Each row corresponds to a different subcommunity and each column represents 
a particular value of \emph{q}.
}
\description{
\code{diversity()} calculates the diversity of a  
supercommunity or series of subcommunities, inclusive of pair-wise 
similarity (whether naive, taxonomic, phenotypic, genetic, phylogenetic, 
functional, \emph{etc.}) between individuals.
}
\details{
There are three common ways to input the \code{measure} argument:

\enumerate{
\item As a single function, \emph{e.g.} \code{diversity(subcommunity.alpha.bar, 0:2)}
\tabular{ll}{
\code{subcommunity.alpha} \tab - estimate of naive-community supercommunity 
diversity \cr
\code{subcommunity.alpha.bar} \tab - similarity-sensitive diversity of 
subcommunity \emph{j} in isolation \cr
\code{subcommunity.rho} \tab - redundancy of subcommunity \emph{j} \cr
\code{subcommunity.rho.bar} \tab - representativeness of subcommunity 
\emph{j} \cr
\code{subcommunity.beta} \tab - distinctiveness of subcommunity \emph{j} \cr
\code{subcommunity.beta.bar} \tab - estimate of effective number of distinct 
subcommunities \cr
\code{subcommunity.gamma} \tab - contribution per individual toward 
supercommunity diversity \cr
\code{supercommunity.A} \tab - naive-community supercommunity diversity \cr
\code{supercommunity.A.bar} \tab - average similarity-sensitive diversity of 
subcommunities \cr
\code{supercommunity.R} for \tab - average redundancy of subcommunities \cr
\code{supercommunity.R.bar} \tab - average representativeness of 
subcommunities \cr
\code{supercommunity.B} \tab - average distinctiveness of subcommunities \cr
\code{supercommunity.B.bar} \tab - effective number of distinct 
subcommunities \cr
\code{supercommunity.G} \tab - supercommunity similarity-sensitive 
diversity \cr
}
\item As a list of functions, \emph{e.g.} 
\code{diversity(c(subcommunity.alpha.bar, supercommunity.R), 0:2)} 
\item As a string, \emph{e.g.} \code{diversity("subcommunity", 0:2)}
\tabular{ll}{
\code{"subMeasures"} \tab - all subcommunity measures of similarity-sensitive 
diversity \cr
\code{"superMeasures"} \tab - all supercommunity measures of 
similarity-sensitive 
diversity \cr
\code{"allMeasures"} \tab - all subcommunity and supercommunity measures of 
similarity-sensitive diversity \cr
}
}
}
\examples{
# Create population
population <- data.frame(A = sample(1:50, 5, replace = TRUE),
                         B = sample(1:50, 5, replace = TRUE))

# Coerse object into a supercommunity
data <- as.supercommunity(population)

# Calculate diversity
# Single measure
output <- diversity(subcommunity.alpha.bar, data, 0:2)

# A number of specific measures 
output <- diversity(c(subcommunity.alpha.bar, supercommunity.G), data, 0:2)
names(output)
output

}

