% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.to.additive.R
\name{diversity.to.additive}
\alias{diversity.to.additive}
\title{diversity.to.additive}
\usage{
diversity.to.additive(diversities)
}
\arguments{
\item{diversities}{Dataframe comprising diversity values, whereby each row 
and column corresponds to a particular subcommunity and q value, 
respectively; q values should be input in column names as character 
strings, e.g. "q1"}
}
\value{
Dataframe comprising 'additive diversity' values, whereby each row
and column corresponds to a particular subcommunity and q value, respectively
}
\description{
Converts diversity values (for any q) to 'additive diversities'.
}

