% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.R
\name{get_label}
\alias{get_label}
\title{get_label}
\usage{
get_label(res, qs = "q")
}
\arguments{
\item{res}{object of class \code{diversity}; output of functions
\code{subdiv()}, \code{metadiv()}, or any of the specific subcommunity- or 
metacommunity-level diversity functions.}

\item{qs}{parameter of conservatism; only required for \code{plot_single()}.}
}
\description{
get_label
}
