% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_title.R
\name{get_title}
\alias{get_title}
\title{Get title}
\usage{
get_title(community_type, measure, symbol = FALSE)
}
\arguments{
\item{community_type}{object of class \code{character}}

\item{measure}{object of class \code{character}}

\item{symbol}{(optional) by default, output is e.g. "Metacommunity x"; if 
symbol is set to \code{TRUE}, output will be given as "x}
}
\description{
Get title
}
\examples{
get_title("subcommunity", "normalised alpha")
get_title("metacommunity", "normalised alpha")
get_title("subcommunity", "normalised alpha", TRUE)
get_title("metacommunity", "normalised alpha", TRUE)

}
