% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.zmatrix.R
\name{hs.names}
\alias{hs.names}
\title{Historic Species}
\usage{
hs.names(tree, node)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{node}{integer corresponding to the node of interest}
}
\value{
object of class \code{character}
}
\description{
This function constructs unique identifiers for all historic species 
ancestral to a given node. Unique identifiers take the form 
\emph{(pds, node-tip)}, where \emph{pds} corresponds to the index associated 
with the present day species descendant, and node-tip corresponds to the 
node index and tip index associated with the historic species itself.
}

