% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_abundance.R
\name{hs_abundance}
\alias{hs_abundance}
\title{hs_abundance}
\usage{
hs_abundance(partition, historic.species, terminal.taxa, Tbar, tag)
}
\arguments{
\item{partition}{\code{vector} or \code{matrix} containing the
relative abundances of individuals or types in their subcommunities. In the
case of phylogenetic metacommunities, these are the relative abundances of
the tips of the tree (the present day species).}

\item{historic.species}{f}

\item{terminal.taxa}{f}

\item{Tbar}{f}

\item{tag}{f}
}
\description{
Helper function that calculates the relative abundance of historic species
for \code{metacommunity()} and \code{chainsaw()}.
}

