% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_parameters.R
\name{hs_parameters}
\alias{hs_parameters}
\title{hs_parameters}
\usage{
hs_parameters(tree)
}
\arguments{
\item{tree}{object of class \code{phylo}}
}
\value{
Returns parameters associated with each historic species
}
\description{
hs_parameters
}
