% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit_params.R
\name{inherit_params}
\alias{inherit_params}
\title{Function arguments}
\arguments{
\item{populations}{object of class \code{collection}}

\item{qs}{\code{vector} of \emph{q} values}
}
\description{
Function arguments
}

