% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-initDiv.R
\docType{class}
\name{initDiv-class}
\alias{initDiv}
\alias{initDiv-class}
\alias{show,initDiv-method}
\title{Class 'initDiv'}
\usage{
\S4method{show}{initDiv}(object)
}
\arguments{
\item{object}{object of class \linkS4class{initDiv}}
}
\description{
Define S4 class \linkS4class{initDiv}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Prints pmatrix
}}
\section{Slots}{

\describe{
\item{\code{measure}}{object of class \code{matrix};}
}}

