% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadiv.R
\docType{methods}
\name{metadiv}
\alias{metadiv}
\alias{metadiv,metacommunity-method}
\alias{metadiv,powermean-method}
\alias{metadiv,relativeentropy-method}
\title{Calculate metacommunity diversity}
\usage{
metadiv(data, qs)

\S4method{metadiv}{powermean}(data, qs)

\S4method{metadiv}{relativeentropy}(data, qs)

\S4method{metadiv}{metacommunity}(data, qs)
}
\arguments{
\item{data}{two-dimensional \code{matrix} of mode \code{numeric}; diversity 
components.}

\item{qs}{\code{vector} of mode \code{numeric}; parameter of conservatism.}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} metacommunity), and \code{measure} (alpha, beta, rho, or gamma).
}
\description{
Generic function for calculating metacommunity diversity.
}
\details{
\code{data} may be input as three different classes:
\itemize{
\item{\code{powermean} calculates raw or normalised supercomunity alpha, rho
or gamma diversity by taking the powermean of diversity components}
\item{\code{relativeentropy} calculates raw or normalised metacommunity beta
diversity by taking the relative entropy of diversity components}
\item{\code{metacommunity} calculates all metacommunity measures of diversity}
}
}
\examples{
pop <- data.frame(a = c(1,3), b = c(1,1))
pop <- pop / sum(pop)
super <- metacommunity(pop)

# Calculate metacommunity gamma diversity (takes the power mean)
g <- raw.gamma(super)
metadiv(g, 0:2)

# Calculate metacommunity beta diversity (takes the relative entropy)
b <- raw.beta(super)
metadiv(b, 0:2)

# Calculate all measures of metacommunity diversity
metadiv(super, 0:2)

}

