% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{norm_alpha}
\alias{norm_alpha}
\title{Low level diversity components: normalised alpha}
\usage{
norm_alpha(meta)
}
\arguments{
\item{meta}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating
normalised alpha diversity.
}
\details{
Values generated from \code{norm_alpha()} may be input into \code{subdiv()}
and \code{metadiv()} to calculate normalised subcommunity/metacommunity
alpha diversity.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate normalised alpha component
a <- norm_alpha(meta)

}

