% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{norm_beta}
\alias{norm_beta}
\title{Low level diversity components: normalised beta}
\usage{
norm_beta(meta)
}
\arguments{
\item{meta}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{relativeentropy}.
}
\description{
Calculates the low-level diversity component necessary for calculating
normalised beta diversity.
}
\details{
Values generated from \code{norm_beta()} may be input into \code{subdiv()} and
\code{metadiv()} to calculate normalised subcommunity/metacommunity beta
diversity.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate normalised beta component
a <- norm_beta(meta)

}
