% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{norm_sub_beta}
\alias{norm_sub_beta}
\title{Similarity-sensitive normalised subcommunity beta diversity}
\usage{
norm_sub_beta(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:
\code{partition} (label attributed to partition), \code{q} (parameter of
conservatism), \code{diversity}, \code{community} (level of diversity,
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure}
(alpha, beta, rho, or gamma).
}
\description{
\code{norm_sub_beta()} is used to calculate an
estimate of the effective number of distinct subcommunities.
}
\details{
\code{norm_sub_beta()} calculates the normalised subcommunity beta
diversity of a series of columns representing independent subcommunity
counts relative to the metacommunity as a whole (by default the sum of the
subcommunities). This measure may be calculated for a series of orders,
repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate normalised subcommunity beta diversity
norm_sub_beta(meta, 0:2)

}
