% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{normalised.alpha}
\alias{normalised.alpha}
\title{Low level diversity components: normalised alpha}
\usage{
normalised.alpha(super)
}
\arguments{
\item{super}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating 
normalised alpha diversity.
}
\details{
Values generated from \code{normalised.alpha()} may be input into \code{subdiv()} 
and \code{superdiv()} to calculate normalised subcommunity/metacommunity 
alpha diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate normalised alpha component
a <- normalised.alpha(super)

}

