% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{normalised.beta}
\alias{normalised.beta}
\title{Low level diversity components: normalised beta}
\usage{
normalised.beta(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns an object of class \code{relativeentropy}.
}
\description{
Calculates the low-level diversity component necessary for calculating 
normalised beta diversity.
}
\details{
Values generated from \code{normalised.beta()} may be input into \code{subdiv()} and
\code{superdiv()} to calculate normalised subcommunity/supercommunity beta 
diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate normalised beta component
a <- normalised.beta(super)

}

