% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{normalised.metacommunity.rho}
\alias{normalised.metacommunity.rho}
\title{Similarity-sensitive normalised metacommunity rho diversity}
\usage{
normalised.metacommunity.rho(super, qs)
}
\arguments{
\item{super}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{normalised.metacommunity.rho()} is used to calculate the 
average representativeness of subcommunities.
}
\details{
\code{normalised.metacommunity.rho()} calculates the total normalised 
metacommunity rho diversity of a series of columns representing independent 
subcommunity counts relative to the metacommunity as a whole (by default 
the sum of the subcommunities). This measure may be calculated for a series 
of orders, repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate normalised metacommunity rho diversity
r <- normalised.rho(super)
metadiv(r, 0:2)
normalised.metacommunity.rho(super, 0:2)

}

