% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{normalised.rho}
\alias{normalised.rho}
\title{Low level diversity components: normalised rho}
\usage{
normalised.rho(super)
}
\arguments{
\item{super}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating 
normalised rho diversity.
}
\details{
Values generated from \code{normalised.rho()} may be input into \code{subdiv()} and
\code{metadiv()} to calculate normalised subcommunity/metacommunity rho 
diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate normalised rho component
a <- normalised.rho(super)

}

