% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{normalised.subcommunity.rho}
\alias{normalised.subcommunity.rho}
\title{Similarity-sensitive normalised subcommunity rho diversity}
\usage{
normalised.subcommunity.rho(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{normalised.subcommunity.rho()} is used to calculate the 
representativeness of subcommunity \emph{j}.
}
\details{
\code{normalised.subcommunity.rho()}, the inverse of \code{subcommunity.beta}), 
calculates the normalised subcommunity rho diversity of a series of columns 
representing independent subcommunities counts relative to a total 
metacommunity (by default the sum of the subcommunities), for a series of 
orders, repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate normalised subcommunity rho diversity
r <- normalised.rho(meta)
subdiv(r, 0:2)
normalised.subcommunity.rho(meta, 0:2)

}

