% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{normalised_alpha}
\alias{normalised_alpha}
\title{Low level diversity components: normalised alpha}
\usage{
normalised_alpha(meta)
}
\arguments{
\item{meta}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating 
normalised alpha diversity.
}
\details{
Values generated from \code{normalised_alpha()} may be input into \code{subdiv()} 
and \code{metadiv()} to calculate normalised subcommunity/metacommunity 
alpha diversity.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate normalised alpha component
a <- normalised_alpha(meta)

}

