% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{normalised_sub_alpha}
\alias{normalised_sub_alpha}
\title{Similarity-sensitive normalised subcommunity alpha}
\usage{
normalised_sub_alpha(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{normalised_sub_alpha()} is used to calculate the 
similarity-sensitive diversity of subcommunity \emph{j} in isolation.
}
\details{
\code{normalised_sub_alpha()} calculates the normalised subcommunity 
alpha diversity of a series of columns representing independent subcommunity 
counts relative to the metacommunity as a whole (by default the sum of the 
subcommunities). This measure may be calculated for a series of orders, 
repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate normalised subcommunity alpha diversity
normalised_sub_alpha(meta, 0:2)

}

