% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2branch.R
\name{phy2branch}
\alias{phy2branch}
\title{Phylogenetic similarity}
\usage{
phy2branch(tree, partition, depth = 1)
}
\arguments{
\item{tree}{object of class \code{phylo}.}

\item{partition}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types (terminal taxa), columns as subcommunities, and each
element containing the relative abundance of types in each subcommunity
relative to the metacommunity as a whole.}

\item{depth}{proportion of total tree height to be conserved (taken as
a proportion from the heighest tip). Describes how much evolutionary history
should be retained, with 0 marking the date of the most recent tip, and 1
(the default) marking the most recent common ancestor. Numbers greater than
1 extend the root of the tree.}
}
\value{
\code{phy2branch()} returns an object of class \code{distance}.
}
\description{
Packages phylo object and depth into an object of class \code{similarity}.
}
\examples{
tree <- ape::rtree(5)
tree$tip.label <- paste0("sp", 1:5)

partition <- matrix(rep(1,10), nrow = 5)
row.names(partition) <- paste0("sp", 1:5)
partition <- partition / sum(partition)

phy2branch(tree, partition)

}
