% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2dist.R
\name{phy2dist}
\alias{phy2dist}
\title{phy2dist}
\usage{
phy2dist(phylo)
}
\arguments{
\item{phylo}{object of class \code{phylo}.}
}
\value{
\code{phy2sim(x)} returns a matrix of pairwise distances.
}
\description{
Converts any phylo object to a matrix of pairwise tip-to-tip distances.
}
\examples{
tree <- ape::rtree(5)
phy2dist(tree)

}
