% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-phy_struct.R
\docType{class}
\name{phy_struct-class}
\alias{phy_struct-class}
\title{phy_struct-class}
\value{
An object of class \code{phy_struct}.
}
\description{
Container for various parameters associated with calculating phylogenetic 
diversity.
}
\section{Fields}{

\describe{
\item{\code{structure}}{\code{matrix} object}

\item{\code{parameters}}{object of class \code{tbl_df}}
}}

