% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_struct.R
\name{phy_struct}
\alias{phy_struct}
\title{Calculate phylogenetic structure matrix}
\usage{
phy_struct(tree)
}
\arguments{
\item{tree}{object of class \code{phylo}}
}
\value{
Returns a \code{list} containing:
\tabular{ll}{
\code{$structure} \tab - each row denotes historic species, columns denote 
terminal taxa, and elements contain branch lengths \cr
\code{$parameters} \tab - information associated with each historic species \cr
\code{$tree} \tab - object of class \code{phylo} \cr
}

Additional slots (accessed with @, not $) include:

\tabular{ll}{
\code{historic.species} \tab - a data_frame containing components associated 
with historic species \cr
\code{terminal_taxa} \tab - na data_frame containing components associated 
with terminal taxa \cr
\code{Tbar} \tab - \cr
\code{structure} \tab -  \cr
}
}
\description{
Converts an object into class \code{phylo} into class \code{phy_struct}.
}
\examples{
tree <- ape::rtree(n = 5)
tree$tip.label <- paste0("sp", seq_along(tree$tip.label))

res <- phy_struct(tree)

}
