% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.dat.R
\name{phylo.dat}
\alias{phylo.dat}
\title{Phylogenetic data}
\usage{
phylo.dat(tree, pds.abundance = rep(1/length(tree$tip.label),
  length(tree$tip.label)))
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{pds.abundance}{\code{vector} of length \eqn{S}; proportional
abundance of present-day species}
}
\value{
S4 generic of class \linkS4class{rdphylo}
}
\description{
Extract historic species names, ancestral and descendant nodes, descendant
present-day species, and proportional abundance. Output as object
}

