% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.pmatrix.R
\name{phylo.pmatrix}
\alias{phylo.pmatrix}
\title{Phylogenetic population matrix}
\usage{
phylo.pmatrix(tree, lookup)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{lookup}{\eqn{hS x 2} matrix; colnames = c('edge', 'subcommunity)}
}
\value{
\eqn{hS x hS} matrix; proportional abundance of historic species
}
\description{
Proportional abundance of ancestral specices; PI(i,b) = (L(b)/Tbar)*pi
}

