% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdplot.R
\docType{methods}
\name{plot.diversity}
\alias{plot.diversity}
\alias{rdplot}
\alias{rdplot,rdiv-method}
\title{rdplot()}
\usage{
rdplot(data, ...)

\S4method{rdplot}{rdiv}(data, style = "normal")
}
\arguments{
\item{data}{object of class \code{rdiv} returns a diversity profile;}

\item{...}{additional parameters}

\item{style}{optional argument; takes \code{'normal'} as default, generating
a standard plot; and \code{'big'} increases font size and line thickness}
}
\value{
object of class \linkS4class{rdiv} returns a diversity profile;
}
\description{
S4 generic function \code{rdplot()}.
}
\examples{
# Species counts

}

