% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot diversity}
\usage{
plot_diversity(res)
}
\arguments{
\item{res}{object of class \code{diversity}; output of functions
\code{metadiv()}, \code{subdiv()}, \code{inddiv} or any of the specific 
subcommunity- or metacommunity-level diversity functions.}
}
\description{
Simple function to plot diversity profiles.
}
\examples{
\dontrun{
# Define metacommunity
pop1 <- data.frame(a = c(1,3), b = c(1,1))
row.names(pop1) <- paste0("sp", 1:2)
pop1 <- pop1/sum(pop1)
meta1 <- metacommunity(pop1)
qs <- c(seq(0,1,.1),2:10, seq(20,100,10),Inf)

# Plot metacommunity beta diversity
b <- raw_beta(meta1)
mc <- metadiv(b, qs)
plot(mc)

# Plot subcommunity beta diversity
sc <- subdiv(b, qs)
plot(sc)

# Plot beta component
ind <- inddiv(b, qs)
plot(ind)

# Plot subcommunity and metacommunity beta diversity
res <- rdiv(list(sc, mc))
plot(res)

# Plot all subcommunity diversity measures
all_sc <- subdiv(meta1, qs)
plot(all_sc)

# Plot all metacommunity diversity measures
all_mc <- metadiv(meta1, qs)
plot(all_mc)

# Plot all diversity measures
all_res <- rdiv(list(all_sc, all_mc))
plot(all_res)

# Try a single population
pop2 <- c(1,3,4)
pop2 <- pop2/sum(pop2)
meta2 <- metacommunity(pop2)
sc <- sub_gamma(meta2, qs)
plot(sc)
mc <- meta_gamma(meta2, qs)
plot(mc)
}

}
