% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inddiv.R
\name{plot_inddiv}
\alias{plot_inddiv}
\title{plot_inddiv}
\usage{
plot_inddiv(res)
}
\arguments{
\item{res}{object of class \code{diversity}; output of functions
\code{subdiv()}, \code{metadiv()}, or any of the specific subcommunity- or 
metacommunity-level diversity functions.}
}
\description{
plot_inddiv
}
\examples{
# Define metacommunity
pop1 <- data.frame(a = c(1,3), b = c(1,1))
row.names(pop1) <- paste0("sp", 1:2)
pop1 <- pop1/sum(pop1)
meta1 <- metacommunity(pop1)
qs <- c(seq(0,1,.1),2:10, seq(20,100,10),Inf)

# Plot metacommunity beta diversity
b <- raw_beta(meta1)
ind <- inddiv(b, qs)
plot(ind)

g <- raw_gamma(meta1)
ind <- inddiv(g, qs)
plot(ind)

all <- inddiv(meta1, qs)
plot(all)

}
