% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single.R
\name{plot_single}
\alias{plot_single}
\title{plot_single}
\usage{
plot_single(res)
}
\arguments{
\item{res}{object of class \code{diversity}; output of functions
\code{subdiv()}, \code{metadiv()}, or any of the specific subcommunity- or 
metacommunity-level diversity functions.}
}
\description{
plot_single
}
