% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-powermean.R
\docType{class}
\name{powermean-class}
\alias{powermean-class}
\title{powermean-class}
\description{
Container for
}
\section{Fields}{

\describe{
\item{\code{.Data}}{two-dimensional \code{matrix} of mode \code{numeric}; contains 
diversity term values}

\item{\code{measure}}{object of class \code{character}; contains an identifier 
associated with the diversity term being calculated}

\item{\code{type_abundance}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains proportional abundance of \emph{types} in the subcommunity 
as a fraction of the supercommunity as a whole (in the phylogenetic case, 
this corresponds to the proportional abundance of historic species, which
is calculated from the proportional abundance of present day species)}

\item{\code{ordinariness}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains ordinariness of types}

\item{\code{subcommunity_weights}}{\code{vector} of mode \code{numeric}; contains
subcommunity weights}

\item{\code{type_weights}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains weight of types within a subcommunity}
}}

