% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.powermean.R
\docType{methods}
\name{powermean}
\alias{as.powermean}
\alias{is.powermean}
\alias{powermean}
\alias{show,powermean-method}
\title{Coerse to powermean}
\usage{
powermean(results, super, tag)

as.powermean(results, super, tag)

is.powermean(x)

\S4method{show}{powermean}(object)
}
\arguments{
\item{results}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains values calculated from diversity-term functions, e.g. 
\code{alphabar()}, \code{beta}, etc.}

\item{super}{object of class \code{supercommunity}; contains proportional 
abundance of types, pair-wise similarity, and other associated variables.}

\item{tag}{object of class \code{character}; contains an identifier 
associated with the calculated diversity-term, e.g. "alphabar", "beta", etc.}

\item{x}{any R object}

\item{object}{object of class \code{powermean}}
}
\value{
\code{powermean(x)} returns an object of class \code{powermean}.

\code{as.powermean(x)} returns an object of class \code{powermean}.

\code{is.powermean(x)} returns TRUE if its argument is a 
powermean, FALSE otherwise.

\code{print(x)} prints an object object of class \code{powermean}
}
\description{
Functions to check if an object is a \code{powermean} or coerce an  
object into a \code{powermean}.
}
\details{
\code{powermean} is a class of object output from calculating 
\code{alpha}, \code{alphabar}, \code{rho}, 
\code{rhobar}, and \code{gamma} diversity terms; 
\code{beta} and \code{betabar} are output as 
\code{relativeentropy} objects.

To calculate the subcommunity or supercommunity diversity, an object of
class \code{powermean} is input to \code{subdiv()} or \code{superdiv()}, 
respectively.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate subcommunity alpha diversity (takes the powermean)
a <- alpha(super)
class(a)

}

