% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qDZ.single.R
\name{qDZ.single}
\alias{qDZ.single}
\title{Similarity-sensitive diversity of a single population}
\usage{
qDZ.single(proportions, q, Z = diag(nrow(proportions)), Zp = Z \%*\%
  proportions)
}
\arguments{
\item{proportions}{Relative proportions of different individuals / types 
in population}

\item{q}{- order of diversity measurement}

\item{Z}{- similarity matrix}

\item{Zp}{- ordinariness of individuals / types in population}
}
\value{
Similarity-sensitive diversity of order q
}
\description{
Calculates the similarity-sensitive diversity of order q of a population 
with given relative proportions.
}
\details{

}
\examples{

}

