% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw.alpha}
\alias{raw.alpha}
\title{Low level diversity components: alpha}
\usage{
raw.alpha(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating alpha
diversity.
}
\details{
Values generated from \code{raw.alpha()} may be input into \code{subdiv()} and
\code{superdiv()} to calculate raw subcommunity/supercommunity alpha diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate raw alpha component
a <- raw.alpha(super)

}

