% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw.beta}
\alias{raw.beta}
\title{Low level diversity components: raw beta}
\usage{
raw.beta(super)
}
\arguments{
\item{super}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{relativeentropy}.
}
\description{
Calculates the low-level diversity component necessary for calculating raw beta
diversity.
}
\details{
Values generated from \code{raw.beta()} may be input into \code{subdiv()} and
\code{superdiv()} to calculate raw subcommunity/metacommunity beta diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate raw beta component
a <- raw.beta(super)

}

