% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw.gamma}
\alias{raw.gamma}
\title{Low level diversity components: gamma}
\usage{
raw.gamma(super)
}
\arguments{
\item{super}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating gamma
diversity.
}
\details{
Values generated from \code{raw.gamma()} may be input into \code{subdiv()} and
\code{superdiv()} to calculate subcommunity/metacommunity gamma diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate gamma component
a <- raw.gamma(super)

}

