% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.metacommunity.alpha}
\alias{raw.metacommunity.alpha}
\title{Similarity-sensitive raw metacommunity alpha diversity}
\usage{
raw.metacommunity.alpha(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{raw.metacommunity.alpha()} is used to calculate the 
naive-community metacommunity diversity.
}
\details{
\code{raw.metacommunity.alpha()} calculates the total metacommunity alpha  
diversity of a series of columns representing independent subcommunity 
counts relative to the metacommunity as a whole (by default the sum of the 
subcommunities). This measure may be calculated for a series of orders, 
repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate raw metacommunity alpha diversity
a <- raw.alpha(meta)
metadiv(a, 0:2)
raw.metacommunity.alpha(meta, 0:2)

}

