% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.metacommunity.beta}
\alias{raw.metacommunity.beta}
\title{Similarity-sensitive raw metacommunity beta diversity}
\usage{
raw.metacommunity.beta(super, qs)
}
\arguments{
\item{super}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{raw.metacommunity.beta()} is used to calculate the 
average distinctiveness of subcommunities.
}
\details{
\code{raw.metacommunity.beta()} calculates the total
metacommunity beta diversity of a series of columns representing
independent subcommunity counts relative to the metacommunity as a
whole (by default the sum of the subcommunities). This measure may
be calculated for a series of orders, repesented as a vector of
\code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate raw metacommunity beta diversity
b <- raw.beta(super)
metadiv(b, 0:2)
raw.metacommunity.beta(super, 0:2)

}

