% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.metacommunity.rho}
\alias{raw.metacommunity.rho}
\title{Similarity-sensitive raw metacommunity rho diversity}
\usage{
raw.metacommunity.rho(super, qs)
}
\arguments{
\item{super}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{raw.metacommunity.rho()} is used to calculate the 
average redundancy of subcommunities.
}
\details{
\code{raw.metacommunity.rho()} calculates the total metacommunity
rho diversity of a series of columns representing independent
subcommunity counts relative to the metacommunity as a whole (by
default the sum of the subcommunities). This measure may be
calculated for a series of orders, repesented as a vector of
\code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate metacommunity rho diversity
r <- raw.rho(super)
superdiv(r, 0:2)
raw.metacommunity.rho(super, 0:2)

}

