% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw.rho}
\alias{raw.rho}
\title{Low level diversity components: raw rho}
\usage{
raw.rho(super)
}
\arguments{
\item{super}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating raw rho
diversity.
}
\details{
Values generated from \code{raw.rho()} may be input into \code{subdiv()} and
\code{superdiv()} to calculate raw subcommunity/metacommunity rho diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- metacommunity(pop)

# Calculate raw rho component
a <- raw.rho(super)

}

