% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.subcommunity.alpha}
\alias{raw.subcommunity.alpha}
\title{Similarity-sensitive raw subcommunity alpha}
\usage{
raw.subcommunity.alpha(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{raw.subcommunity.alpha()} is used to calculate an estimate of 
naive-community supercommunity diversity.
}
\details{
\code{raw.subcommunity.alpha()} calculates the subcommunity alpha diversity of 
a series of columns representing independent subcommunity counts relative 
to the supercommunity as a whole (by default the sum of the subcommunities). 
This measure may be calculated for a series of orders, repesented as a  
vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate raw subcommunity alpha diversity
a <- raw.alpha(super)
subdiv(a, 0:2)

}

