% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.subcommunity.beta}
\alias{raw.subcommunity.beta}
\title{Similarity-sensitive raw subcommunity beta diversity}
\usage{
raw.subcommunity.beta(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or supercommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{raw.subcommunity.beta()} is used to calculate the 
distinctiveness of subcommunity \emph{j}.
}
\details{
\code{raw.subcommunity.beta()} calculates the subcommunity beta diversity of a 
series of columns representing independent subcommunity counts relative to 
the supercommunity as a whole (by default the sum of the subcommunities).  
This measure may be calculated for a series of orders, repesented as a  
vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate raw subcommunity beta diversity
b <- raw.beta(super)
subdiv(b, 0:2)
raw.subcommunity.beta(super, 0:2)

}

