% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.supercommunity.beta}
\alias{raw.supercommunity.beta}
\title{Similarity-sensitive raw supercommunity beta diversity}
\usage{
raw.supercommunity.beta(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{raw.supercommunity.beta()} is used to calculate the 
average distinctiveness of subcommunities.
}
\details{
\code{raw.supercommunity.beta()} calculates the total
supercommunity beta diversity of a series of columns representing
independent subcommunity counts relative to the supercommunity as a
whole (by default the sum of the subcommunities). This measure may
be calculated for a series of orders, repesented as a vector of
\code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate raw supercommunity beta diversity
b <- beta(super)
superdiv(b, 0:2)
raw.supercommunity.beta(super, 0:2)

}

