% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.supercommunity.rho}
\alias{raw.supercommunity.rho}
\title{Similarity-sensitive raw supercommunity rho diversity}
\usage{
raw.supercommunity.rho(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or supercommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{raw.supercommunity.rho()} is used to calculate the 
average redundancy of subcommunities.
}
\details{
\code{raw.supercommunity.rho()} calculates the total supercommunity
rho diversity of a series of columns representing independent
subcommunity counts relative to the supercommunity as a whole (by
default the sum of the subcommunities). This measure may be
calculated for a series of orders, repesented as a vector of
\code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate supercommunity rho diversity
r <- raw.rho(super)
superdiv(r, 0:2)
raw.supercommunity.rho(super, 0:2)

}

