% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw_alpha}
\alias{raw_alpha}
\title{Low level diversity components: alpha}
\usage{
raw_alpha(meta)
}
\arguments{
\item{meta}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating alpha
diversity.
}
\details{
Values generated from \code{raw_alpha()} may be input into \code{subdiv()} and
\code{metadiv()} to calculate raw subcommunity/metacommunity alpha diversity.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate raw alpha component
a <- raw_alpha(meta)

}
