% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw_beta}
\alias{raw_beta}
\title{Low level diversity components: raw beta}
\usage{
raw_beta(meta)
}
\arguments{
\item{meta}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{relativeentropy}.
}
\description{
Calculates the low-level diversity component necessary for calculating raw beta
diversity.
}
\details{
Values generated from \code{raw_beta()} may be input into \code{subdiv()} and
\code{metadiv()} to calculate raw subcommunity/metacommunity beta diversity.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate raw beta component
a <- raw_beta(meta)

}
