% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw_meta_rho}
\alias{raw_meta_rho}
\title{Similarity-sensitive raw metacommunity rho diversity}
\usage{
raw_meta_rho(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:
\code{partition} (label attributed to partition), \code{q} (parameter of
conservatism), \code{diversity}, \code{community} (level of diversity,
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure}
(alpha, beta, rho, or gamma).
}
\description{
\code{raw_meta_rho()} is used to calculate the
average redundancy of subcommunities.
}
\details{
\code{raw_meta_rho()} calculates the total metacommunity
rho diversity of a series of columns representing independent
subcommunity counts relative to the metacommunity as a whole (by
default the sum of the subcommunities). This measure may be
calculated for a series of orders, repesented as a vector of
\code{qs}.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate metacommunity rho diversity
raw_meta_rho(meta, 0:2)

}

