% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw_rho}
\alias{raw_rho}
\title{Low level diversity components: raw rho}
\usage{
raw_rho(meta)
}
\arguments{
\item{meta}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating raw rho
diversity.
}
\details{
Values generated from \code{raw_rho()} may be input into \code{subdiv()} and
\code{metadiv()} to calculate raw subcommunity/metacommunity rho diversity.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate raw rho component
a <- raw_rho(meta)

}

