% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw_sub_rho}
\alias{raw_sub_rho}
\title{Similarity-sensitive raw subcommunity rho diversity}
\usage{
raw_sub_rho(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:
\code{partition} (label attributed to partition), \code{q} (parameter of
conservatism), \code{diversity}, \code{community} (level of diversity,
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure}
(alpha, beta, rho, or gamma).
}
\description{
\code{raw_sub_rho()} is used to calculate the redundancy of
subcommunity \emph{j}.
}
\details{
\code{raw_sub_rho()}, the inverse of \code{raw_sub_beta()},
calculates the raw subcommunity rho diversity of a series of columns
representing independent subcommunity counts relative to the
metacommunity as a whole (by default the sum of the subcommunities). This
measure may be calculated for a series of orders, repesented as a vector
of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate raw subcommunity rho diversity
raw_sub_rho(meta, 0:2)

}

