% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdiv.R
\docType{methods}
\name{rdiv}
\alias{rdiv}
\alias{rdiv,data.frame-method}
\alias{rdiv,list-method}
\title{Generate rdiv}
\usage{
rdiv(res)

\S4method{rdiv}{data.frame}(res)

\S4method{rdiv}{list}(res)
}
\arguments{
\item{res}{object of class \code{rdiv}}
}
\value{
\code{rdiv()} returns an object of class \code{rdiv}.
}
\description{
Functions to check if an object is a \code{rdiv}
}
\examples{
pop <- data.frame(a = c(1,3), b = c(1,1))
row.names(pop) <- paste0("sp", 1:2)
pop <- pop/sum(pop)
meta <- metacommunity(pop)
qs <- c(seq(0,1,.1),2:10, seq(20,100,10),Inf)

b <- raw_beta(meta)
sc <- subdiv(b, qs)
mc <- metadiv(b, qs)
class(sc)
class(mc)

res <- rbind.data.frame(sc, mc)
res <- rdiv(res)
class(res)

res <- list(sc, mc)
res <- rdiv(res)
class(res)

}
