% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDiversity.R
\docType{package}
\name{RDiversity}
\alias{RDiversity}
\alias{RDiversity-package}
\title{RDiversity}
\description{
The \code{RDiversity} package calculates the similarity-sensitive diversity 
of a population and it's partitions.
}
\details{
This package calculates, or takes as input, the similarity between 
types; which may be unique, taxonimic, phenotypic, genetic, 
phylogenetic, functional, and so on. A type is defined as the categorical 
variable in which the similarity of individuals is assessed.
}
\author{
Sonia Mitchell <soniamitchell@gmail.com>, Richard Reeve 
\cr
Maintainer: Sonia Mitchell <soniamitchell@gmail.com>
}

