% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-phylo.R
\docType{class}
\name{rdphylo-class}
\alias{rdphylo-class}
\title{Class 'rdphylo'}
\description{
Class 'rdphylo'
}
\section{Slots}{

\describe{
\item{\code{hs.name}}{\code{vector} of length hS; historic species names}

\item{\code{hs.pds}}{\code{vector} of length hS; descendant present day species}

\item{\code{hs.edge}}{\code{matrix}; ancestral and descendant nodes}

\item{\code{hs.abundance}}{\code{vector} of length hS; historic species abundance}
}}

