% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rdphylo.R
\docType{methods}
\name{rdphylo}
\alias{as.rdphylo}
\alias{is.rdphylo}
\alias{rdphylo}
\alias{show,rdphylo-method}
\title{Coerse to rdphylo}
\usage{
rdphylo(tree, pds.abundance = matrix(rep(1/length(tree$tip.label),
  length(tree$tip.label))))

as.rdphylo(tree, pds.abundance = matrix(rep(1/length(tree$tip.label),
  length(tree$tip.label))))

is.rdphylo(x)

\S4method{show}{rdphylo}(object)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{pds.abundance}{\code{vector} of length \eqn{S}; proportional
abundance of present-day species}

\item{x}{any R object}

\item{object}{object of class \code{rdphylo}}
}
\value{
S4 generic of class \code{rdphylo} containing historic 
species: names, ancestral and descendant nodes, descendant present day 
species, and proportional abundance.
}
\description{
Functions to check if an object is a \code{rdphylo} or coerce an  
object into a \code{rdphylo}.
}

