% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdplot.R
\name{rdplot}
\alias{rdplot}
\title{rdplot()}
\usage{
rdplot(data, ...)
}
\arguments{
\item{data}{object of class \code{rdiv} returns a diversity profile; object
of class \linkS4class{rdphylo} returns phylogenetic tree}

\item{...}{additional parameters}
}
\value{
object of class \linkS4class{rdiv} returns a diversity profile; object
of class \linkS4class{rdphylo} returns phylogenetic tree
}
\description{
S4 generic function \code{rdplot()}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace=T),
                        subcommunityB = sample(1:50, 5, replace=T))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

# Calculate diversity
output <- diversity(subcommunity.alpha.bar, data, 0:2)

rdplot(output)

}

