% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.relativeentropy.R
\docType{methods}
\name{relativeentropy}
\alias{as.relativeentropy}
\alias{is.relativeentropy}
\alias{relativeentropy}
\alias{show,relativeentropy-method}
\title{Coerse to relativeentropy}
\usage{
relativeentropy(results, super, tag)

as.relativeentropy(results, super, tag)

is.relativeentropy(x)

\S4method{show}{relativeentropy}(object)
}
\arguments{
\item{results}{two-dimensional \code{matrix} of mode \code{numeric};}

\item{super}{object of class \code{supercommunity};}

\item{tag}{object of class \code{character};}

\item{x}{any R object}

\item{object}{object of class \code{relativeentropy}}
}
\value{
object of class \code{relativeentropy}

\code{is.relativeentropy(x)} returns TRUE if its argument is a 
relativeentropy, FALSE otherwise.
}
\description{
Functions to check if an object is a \code{relativeentropy} or coerce an  
object into a \code{relativeentropy}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate subcommunity beta diversity (takes the relative entropy)
a <- beta(super)
class(a)

}

