% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{rho}
\alias{rho}
\title{Low level diversity components: rho}
\usage{
rho(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns
}
\description{
Calculates the low-level diversity component necessary for calculating rho
diversity.
}
\details{
Values generated from \code{rho()} may be input into \code{subdiv()} and
\code{superdiv} to calcualte raw subcommunity/supercommunity rho diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate rho component
a <- rho(super)

}

