% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{rhobar}
\alias{rhobar}
\title{Low level diversity components}
\usage{
rhobar(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns
}
\description{
Calculates the low-level diversity component necessary for calculating 
normalised rho diversity.
}
\details{
Values generated from \code{rhobar()} may be input into \code{subdiv()} and
\code{superdiv} to calcualte normalised subcommunity/supercommunity rho 
diversity.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate normalised rho component
a <- rhobar(super)

}

