% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_matrix.R
\name{s_matrix}
\alias{s_matrix}
\title{Phylogenetic Similarity Matrix}
\usage{
s_matrix(tree, ps)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{ps}{object of class \code{phy_struct}}
}
\value{
\eqn{hS x hS} matrix; pair-wise similarity of historic species
}
\description{
Function to calculate phylogenetic similarity matrix from a phylogeny.
}
\examples{
tree <- ape::rtree(n = 5)

}
