% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.collection.R
\name{set.collection}
\alias{set.collection}
\title{Set Collection}
\usage{
set.collection(data, similarity = NA, zmatrix = NA, lookup = NA)
}
\arguments{
\item{data}{Object of class \code{matrix}}

\item{similarity}{Object of class \code{character}}

\item{zmatrix}{Object of class \code{matrix}}

\item{lookup}{Object of class \code{data.frame}}
}
\value{
Object of class \code{initDiv}
}
\description{
This object contains
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

class(data)

data

data@zmatrix

}

