% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-similarity.R
\docType{class}
\name{similarity-class}
\alias{similarity-class}
\title{similarity-class}
\description{
Container for class \code{similarity}.
}
\section{Fields}{

\describe{
\item{\code{phylo}}{object of class \code{phylo}.}

\item{\code{depth}}{object of class \code{numeric}.}
}}

