% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Generate similarity object}
\usage{
similarity(phylo, depth = 1)
}
\arguments{
\item{phylo}{object of class \code{phylo}}

\item{depth}{object of class \code{numeric}}
}
\value{
similarity() returns an object of class \code{similarity}
}
\description{
Used by \code{phy2branch} to generate a \code{similarity} object.
}
