% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_phylo.R
\name{similarity_phylo}
\alias{similarity_phylo}
\title{Phylogenetic Similarity Matrix}
\usage{
similarity_phylo(data, pds.abundance)
}
\arguments{
\item{data}{object of class \code{rdphylo}}

\item{pds.abundance}{vector of length {S}; relative abundance of present-day 
species}
}
\value{
\eqn{hS x hS} matrix; pair-wise similarity of historic species
}
\description{
Function to calculate phylogenetic similarity matrix from a phylogeny.
}
\examples{
tree <- ape::rtree(n = 5)
pop <- rep(1/5, 5)
new.tree <- rdphylo(pop, tree)
Z <- similarity_phylo(new.tree)

}

