% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plot.R
\name{simple_plot}
\alias{simple_plot}
\title{Plot diversity}
\usage{
simple_plot(res)
}
\arguments{
\item{res}{object of class \code{diversity}; output of functions
\code{subdiv()}, \code{metadiv()}, or any of the specific subcommunity- or 
metacommunity-level diversity functions.}
}
\description{
Simple function to plot diversity profiles.
}
\examples{
# Define metacommunity
pop1 <- data.frame(a = c(1,3), b = c(1,1))
row.names(pop1) <- paste0("sp", 1:2)
pop1 <- pop1/sum(pop1)
meta1 <- metacommunity(pop1)
qs <- 0:2

# Plot subcommunity beta diversity
b <- raw_beta(meta1)
sc <- subdiv(b, qs)
plot(sc)

# Plot metacommunity beta diversity
mc <- metadiv(b, qs)
plot(mc)

# Plot subcommunity and metacommunity beta diversity
res <- diversity(list(sc, mc))
plot(res)

# Plot all subcommunity diversity measures
all_sc <- subdiv(meta1, qs)
plot(all_sc)

# Plot all metacommunity diversity measures
all_mc <- metadiv(meta1, qs)
plot(all_mc)

# Plot all diversity measures
all_res <- diversity(list(all_sc, all_mc))
plot(all_res)

# Try a single population
pop2 <- c(1,3,4)
pop2 <- pop2/sum(pop2)
meta2 <- metacommunity(pop2)
sc <- sub_gamma(meta2, qs)
plot(sc)
mc <- meta_gamma(meta2, qs)
plot(mc)

# Try large number of subcommunities
pop3 <- matrix(sample(1000), ncol = 100)
row.names(pop3) <- paste0("sp", 1:10)
pop3 <- pop3/sum(pop3)
meta3 <- metacommunity(pop3)
sc <- sub_gamma(meta3, qs)
plot(sc)

}
