% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smatrix.R
\name{smatrix}
\alias{smatrix}
\title{Phylogenetic Similarity Matrix}
\usage{
smatrix(ps)
}
\arguments{
\item{ps}{\code{phy_struct()} output \code{$parameters}.}
}
\value{
\eqn{hS x hS} matrix; pair-wise similarity of historic species
}
\description{
Function to calculate phylogenetic similarity matrix from a phylogeny.
}
\examples{
tree <- ape::rtree(n = 5)
tree$tip.label <- paste0("sp", seq_along(tree$tip.label))
partition <- cbind(a = c(1,1,1,0,0), b = c(0,1,0,1,1))
row.names(partition) <- tree$tip.label
partition <- partition / sum(partition)

ps <- phy_struct(tree)
smatrix(ps)

}
