% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_phylo.R
\name{structure_phylo}
\alias{structure_phylo}
\title{Phylogenetic Similarity Matrix}
\usage{
structure_phylo(tree, historic.species)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{historic.species}{vector of length {S}; relative abundance of present-day 
species}
}
\value{
\eqn{hS x hS} matrix; pair-wise similarity of historic species
}
\description{
Function to calculate phylogenetic similarity matrix from a phylogeny.
}
\examples{
tree <- ape::rtree(n = 5)
new.tree <- rdphylo(tree = tree)
new.tree@structure

}

