% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.alpha.bar.R
\name{subcommunity.alpha.bar}
\alias{subcommunity.alpha.bar}
\title{Similarity-sensitive Normalised subcommunity.alpha}
\usage{
subcommunity.alpha.bar(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}
}
\value{
An array of diversities, first dimension representing 
sub-communities, and last representing values of q
}
\description{
Calculates the diversity of a series of columns representing
independent subcommunity counts, for a series of orders, repesented as
a vector of qs
}
\seealso{
\code{\link{subcommunity.alpha}}
}

