% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.alpha.R
\name{subcommunity.alpha}
\alias{subcommunity.alpha}
\title{Similarity-sensitive Raw subcommunity.alpha}
\usage{
subcommunity.alpha(populations, qs, Z = diag(nrow(populations)),
  normalise = F)
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}

\item{normalise}{Normalise probability distribution to sum to 1}
}
\value{
An array of diversities, first dimension representing 
subcommunities, and last representing values of q
}
\description{
Calculates the diversity of a series of columns representing
independent subcommunity counts, for a series of orders, repesented as
a vector of qs
}
\seealso{
\code{\link{subcommunity.alpha.bar, subcommunity.A, supercommunity.A.bar}}
}

