% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{subcommunity.beta.bar}
\alias{subcommunity.beta.bar}
\title{Similarity-sensitive normalised subcommunity.beta diversity}
\usage{
subcommunity.beta.bar(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{subcommunity.beta.bar()} is used to calculate an 
estimate of the effective number of distinct subcommunities.
}
\details{
\code{subcommunity.beta.bar()} calculates the normalised subcommunity beta 
diversity of a series of columns representing independent subcommunity 
counts relative to the supercommunity as a whole (by default the sum of the 
subcommunities). This measure may be calculated for a series of orders, 
repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate normalised subcommunity beta diversity (takes the relative 
# entropy)
a <- betabar(super)
subdiv(a, 0:2)

}

