% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{subcommunity.beta}
\alias{subcommunity.beta}
\title{Similarity-sensitive raw subcommunity.beta diversity}
\usage{
subcommunity.beta(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{subcommunity.beta()} is used to calculate the 
distinctiveness of subcommunity \emph{j}.
}
\details{
\code{subcommunity.beta()} calculates the subcommunity beta diversity of a 
series of columns representing independent subcommunity counts relative to 
the supercommunity as a whole (by default the sum of the subcommunities).  
This measure may be calculated for a series of orders, repesented as a  
vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate subcommunity beta diversity (takes the relative entropy)
a <- beta(super)
subdiv(a, 0:2)

}

