% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.gamma.bar.R
\name{subcommunity.gamma.bar}
\alias{subcommunity.gamma.bar}
\title{Similarity-sensitive Normalised subcommunity.gamma diversity}
\usage{
subcommunity.gamma.bar(populations, qs, Z = diag(nrow(populations)), ...)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of \emph{q}
}
\description{
Calculates the diversity of a series of columns representing independent 
subcommunities counts relative to a total supercommunity (by default the 
sum of the subcommunities), for a series of orders, repesented as a  
vector of \code{qs}.
}
\seealso{
\code{\link{subcommunity.gamma}}, \code{\link{supercommunity.G}}, \code{\link{supercommunity.G.bar}}
}

