% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{subcommunity.rho.bar}
\alias{subcommunity.rho.bar}
\title{Similarity-sensitive normalised subcommunity.rho diversity}
\usage{
subcommunity.rho.bar(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{subcommunity.rho.bar()} is used to calculate the 
representativeness of subcommunity \emph{j}.
}
\details{
\code{subcommunity.rho.bar()}, the inverse of \code{subcommunity.beta}), 
calculates the normalised subcommunity rho diversity of a series of columns 
representing independent subcommunities counts relative to a total 
supercommunity (by default the sum of the subcommunities), for a series of 
orders, repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate noramlised subcommunity rho diversity (takes the power mean)
a <- rhobar(super)
subdiv(a, 0:2)

}

