% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-supercommunity.R
\docType{class}
\name{supercommunity-class}
\alias{show,supercommunity-method}
\alias{supercommunity}
\alias{supercommunity-class}
\title{Class 'supercommunity'}
\usage{
\S4method{show}{supercommunity}(object)
}
\arguments{
\item{object}{object of class \linkS4class{supercommunity}}
}
\description{
Container for proportional abundance and similarity matrices.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: prints pmatrix
}}
\section{Fields}{

\describe{
\item{\code{.Data}}{\code{matrix} of mode \code{numeric}; proportional abundance 
of \emph{types}}

\item{\code{zmatrix}}{\code{matrix} of mode \code{numeric}; pairwise similarity 
between \emph{types}}
}}

