% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.supercommunity.R
\docType{methods}
\name{supercommunity}
\alias{ANY-method}
\alias{as.supercommunity}
\alias{supercommunity}
\alias{supercommunity,}
\alias{supercommunity,data.frame,matrix-method}
\alias{supercommunity,data.frame,missing-method}
\alias{supercommunity,integer,matrix-method}
\alias{supercommunity,integer,missing-method}
\alias{supercommunity,matrix,matrix-method}
\alias{supercommunity,matrix,missing-method}
\alias{supercommunity,phylo,matrix-method}
\alias{supercommunity,phylo,missing-method}
\title{Coerse to Supercommunity}
\usage{
supercommunity(partition, similarity, ...)

\S4method{supercommunity}{data.frame,missing}(partition, similarity = NA)

\S4method{supercommunity}{integer,missing}(partition, similarity = NA)

\S4method{supercommunity}{data.frame,matrix}(partition, similarity = NA)

\S4method{supercommunity}{integer,matrix}(partition, similarity = NA)

\S4method{supercommunity}{matrix,missing}(partition, similarity = NA)

\S4method{supercommunity}{matrix,matrix}(partition, similarity = NA)

\S4method{supercommunity}{phylo,missing}(partition, similarity,
  pds.abundance = NA)

\S4method{supercommunity}{phylo,matrix}(partition, similarity,
  pds.abundance = NA)

supercommunity(partition, similarity, ...)
}
\arguments{
\item{partition}{\code{matrix} (usually two-dimensional) of mode numeric;
relative abundance of types}

\item{similarity}{(optional) two-dimensional \code{matrix} of mode numeric; 
pair-wise similarity of types. Default sets similarity to the naive-type 
case, where types are completely distinct.}

\item{...}{additional arguments.}

\item{pds.abundance}{\code{vector} of mode \code{numeric}; containing the 
proportional abundance of present day species (leaves)}
}
\value{
Returns an object of class \code{supercommunity}; an S4 object 
containing five slots (see Details).
}
\description{
Functions to check if an object is a \code{supercommunity} or coerce an  
object into a \code{supercommunity}.
}
\details{
\enumerate{
\item .Data (partition) - proportional abundance of samples (usually types, 
except in the phylogenetic case where samples correspond to the present day 
species)
\item similarity - pairwise similarity between \emph{types}
\item type_abundance - proportional abundance of \emph{types} in the  
subcommunity as a fraction of the supercommunity as a whole (in the  
phylogenetic case, this corresponds to the proportional abundance of 
historic species, which is calculated from the proportional abundance of 
present day species)
\item ordinariness - ordinariness of types
\item subcommunity_weights - subcommunity weights
\item type_weights - weight of types within a subcommunity
}
}
\examples{
tree <- ape::rtree(n = 5)
a <- supercommunity(tree)
str(a)
a

}
\seealso{
\code{\link{supercommunity-class}}
}

