% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{supercommunity.G}
\alias{supercommunity.G}
\title{Similarity-sensitive raw supercommunity.G diversity}
\usage{
supercommunity.G(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{supercommunity.G()} is used to calculate the 
supercommunity similarity-sensitive diversity.
}
\details{
\code{supercommunity.G()} calculates the total supercommunity gamma diversity 
of a series of columns representing independent subcommunity counts relative 
to the supercommunity as a whole (by default the sum of the subcommunities).  
This measure may be calculated for a series of orders, repesented as a vector 
of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate supercommunity gamma diversity (takes the power mean)
a <- gamma(super)
superdiv(a, 0:2)

}

