% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.R.R
\name{supercommunity.R}
\alias{supercommunity.R}
\title{Similarity-sensitive Raw supercommunity.R diversity}
\usage{
supercommunity.R(populations, qs, Z = diag(nrow(populations)),
  normalise = F)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity.R diversity of a series of 
columns representing subcommunity counts, for a series of orders, 
repesented as a vector of \code{qs}.
}
\seealso{
\code{\link{supercommunity.R.bar}}, \code{\link{subcommunity.rho}}, \code{\link{subcommunity.rho.bar}}
}

