% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.zmatrix.R
\name{tax.zmatrix}
\alias{tax.zmatrix}
\title{Taxonomic similarity matrix}
\usage{
tax.zmatrix(data, lookup)
}
\arguments{
\item{data}{\eqn{S x N} \code{matrix}; population counts}

\item{lookup}{\code{data.frame} with colnames = c('Species', 'Genus', 'Family', 'Subclass')}
}
\value{
Returns an \eqn{S x S} \code{matrix}; pair-wise taxonomic similarity
}
\description{
This function calculates taxonomic similarity based on Shimatani's 
index of taxonomic similarity (See Details).
}
\details{
Shimatani's taxonomic similarity index is defined:
\tabular{ll}{
\code{species.similarity} \tab 1 \cr
\code{genus.similarity} \tab 0.75 \cr
\code{family.similarity} \tab 0.5 \cr
\code{subclass.similarity} \tab 0.25 \cr
\code{other.similarity} \tab 0 \cr
}
}

