% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxfac.R
\name{taxfac}
\alias{taxfac}
\title{taxfac}
\usage{
taxfac(lookup)
}
\arguments{
\item{lookup}{\code{data.frame} with colnames corresponding to nested 
hierarchical levels, e.g. c('Species', 'Genus', 'Family', 'Subclass')}
}
\description{
taxfac
}
