% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxid.R
\name{taxid}
\alias{taxid}
\title{taxid}
\usage{
taxid(lookup)
}
\arguments{
\item{lookup}{\code{data.frame} with colnames corresponding to nested 
hierarchical levels, e.g. c('Species', 'Genus', 'Family', 'Subclass')}
}
\description{
Generate taxanomic codes for each species by converting species, genus, 
family, and subclass into factors
}
