% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmask.R
\name{taxmask}
\alias{taxmask}
\title{taxmask}
\usage{
taxmask(lookup)
}
\arguments{
\item{lookup}{Lookup table}
}
\description{
taxmask
}
\examples{
\dontrun{
# Create Lookup table
Species <- c("tenuifolium", "asterolepis", "simplex var.grandiflora", "simplex var.ochnacea")
Genus <- c("Protium", "Quararibea", "Swartzia", "Swartzia")
Family <- c("Burseraceae", "Bombacaceae", "Fabaceae", "Fabaceae")
Subclass <- c("Sapindales", "Malvales", "Fabales", "Fabales")
lookup <- cbind.data.frame(Species, Genus, Family, Subclass)

taxmask(lookup)
}

}
