% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxvec.R
\name{taxvec}
\alias{taxvec}
\title{taxvec}
\usage{
taxvec(similarity, row)
}
\arguments{
\item{similarity}{khj}

\item{row}{jkh}
}
\description{
Find the taxonomic similarity of a single species to all other species
}
