% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbar.R
\name{tbar}
\alias{tbar}
\title{tbars}
\usage{
tbar(ps, partition)
}
\arguments{
\item{ps}{\code{phy_struct()} output.}

\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with
rows as types, columns as subcommunities, and elements containing relative
abundances of types in subcommunities. In the case of phylogenetic
metacommunities, these are the relative abundances of terminal taxa.}
}
\description{
Function to calculate T_bar.
}
