% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.zmatrix.R
\name{unique.zmatrix}
\alias{unique.zmatrix}
\title{Naive-type similarity matrix}
\usage{
\method{unique}{zmatrix}(data)
}
\arguments{
\item{data}{\eqn{S x N} \code{matrix}; population counts}
}
\value{
Returns an \eqn{S x S} \code{matrix}; identity matrix where each \emph{type} 
is completely distinct
}
\description{
This function returns a naive-type similarity matrix, whereby \emph{types} 
are considered to be completely similar to themselves, but completely 
distinct from each other.
}

