% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrix.R
\name{z_matrix}
\alias{z_matrix}
\title{z_matrix}
\usage{
z_matrix(partition, smatrix, ps)
}
\arguments{
\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with 
rows as types, columns as subcommunities, and elements containing relative 
abundances of types in subcommunities. In the case of phylogenetic 
metacommunities, these are the relative abundances of terminal taxa.}

\item{smatrix}{object}

\item{ps}{object}
}
\value{
\eqn{hS x hS} matrix; pair-wise similarity of historic species
}
\description{
Function to
}
\examples{
tree <- ape::rtree(n = 5)
tree$tip.label <- paste0("sp", seq_along(tree$tip.label))
partition <- cbind(a = c(1,1,1,0,0), b = c(0,1,0,1,1))
row.names(partition) <- tree$tip.label
partition <- partition / sum(partition)
ps <- phy_struct(tree)
smatrix <- s_matrix(ps)

z_matrix(partition, smatrix, ps)

}
