% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{alpha}
\alias{alpha}
\alias{alphabar}
\alias{beta}
\alias{betabar}
\alias{gamma}
\alias{rho}
\alias{rhobar}
\title{Low level diversity components}
\usage{
alpha(super)

alphabar(super)

rho(super)

rhobar(super)

beta(super)

betabar(super)

gamma(super)
}
\arguments{
\item{super}{object of class \code{supercommunity}.}
}
\value{
Returns
}
\description{
Calculates the diversity of a series of columns representing
independent subcommunity counts, for a series of orders, repesented as
a vector of \code{qs}.
}
\details{

}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Coerse object into a supercommunity
data <- supercommunity(population)

# Calculate diversity
subcommunity.alpha(data, 0:2)

}

