% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.zmatrix.R
\name{calculate.zmatrix}
\alias{calculate.zmatrix}
\title{Calculate Similarity Matrix (Z)}
\usage{
calculate.zmatrix(type, data, lookup = NA)
}
\arguments{
\item{type}{character string; defining which similarity measure is being 
called for.}

\item{data}{matrix of count data}

\item{lookup}{required for calculating \code{taxonomic} similarity; contains 
four columns, listing 'Species', 'Genus', 'Family', and 'Subclass'}
}
\value{
matrix of pair-wise similarity indices
}
\description{
The function \code{calculate.zmatrix} calculates the similarity matrix.
}
\details{
The argument \code{type} takes the following inputs: \cr
\tabular{ll}{
\code{unique} \tab - assumes all \emph{types} are unique \cr
\code{taxonomic} \tab - assumes all \emph{types} are species; calculates a 
similarity matrix based on Shimatani's taxonomic similarity \cr
}
}

