% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainsaw.R
\name{chainsaw}
\alias{chainsaw}
\title{Cut phylogeny}
\usage{
chainsaw(data, leaf.abundance, interval)
}
\arguments{
\item{data}{object of class \code{rdphylo}}

\item{leaf.abundance}{proportional abundance of tips}

\item{interval}{proportion of total tree height to be conserved (taken as 
a proportion from the heighest tip)}
}
\value{
Returns an object of class \code{supercommunity}.
}
\description{
Cut phylogeny
}

