% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_partition.R
\name{check_partition}
\alias{check_partition}
\title{Check partition matrix}
\usage{
check_partition(partition)
}
\arguments{
\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with 
rows as types, columns as subcommunities, and elements containing relative 
abundances of types in subcommunities. In the case of phylogenetic 
metacommunities, these are the relative abundances of terminal taxa.}
}
\value{
Returns a two-dimensions \code{matrix} of mode \code{numeric}. If 
the partition matrix was valid, this should be identical to that which was
input as an argument.
}
\description{
\code{check_partition()} is used to validate partition matrices.
}
