% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_partition.R
\name{check_partition}
\alias{check_partition}
\title{Check partition matrix}
\usage{
check_partition(partition)
}
\arguments{
\item{partition}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types (species), columns as subcommunities, and each
element containing the relative abundance of types in each subcommunity
relative to the metacommunity as a whole. In the phylogenetic case, this
corresponds to the proportional abundance of historical species, which is
calculated from the proportional abundance of terminal taxa}
}
\value{
Returns a two-dimensions \code{matrix} of mode \code{numeric}. If
the partition matrix was valid, this should be identical to that which was
input as an argument.
}
\description{
\code{check_partition()} is used to validate partition matrices.
}
