% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phypartition.R
\name{check_phypartition}
\alias{check_phypartition}
\title{check_phypartition}
\usage{
check_phypartition(tip_labels, partition)
}
\arguments{
\item{tip_labels}{vector containing elements of class \code{character}.}

\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with
rows as types, columns as subcommunities, and elements containing relative
abundances of types in subcommunities. In the case of phylogenetic
metacommunities, these are the relative abundances of terminal taxa.}
}
\value{
Returns a two-dimensions \code{matrix} of mode \code{numeric}. If 
the partition matrix was valid, this should be identical to that which was
input as an argument.
}
\description{
\code{check_phypartition()} is used to validate partition matrices for use 
with phylogenies.
}
