% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-collection.R
\docType{class}
\name{collection-class}
\alias{collection}
\alias{collection-class}
\alias{show,collection-method}
\title{Class 'collection'}
\usage{
\S4method{show}{collection}(object)
}
\arguments{
\item{object}{object of class \linkS4class{collection}}
}
\description{
Container for proportional abundance and similarity matrices.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: prints pmatrix
}}
\section{Fields}{

\describe{
\item{\code{.Data}}{\code{matrix} of mode \code{numeric}; proportional abundance 
of \emph{types}}

\item{\code{zmatrix}}{\code{matrix} of mode \code{numeric}; pairwise similarity 
between \emph{types}}
}}

