% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2sim.R
\name{dist2sim}
\alias{dist2sim}
\title{Distance to similarity}
\usage{
dist2sim(dist, transform, k = 1, normalise = TRUE, max_d = 1)
}
\arguments{
\item{dist}{two-dimensional \code{matrix} of mode \code{numeric} with rows as 
types, columns as types, and elements containing the pairwise distance 
between types.}

\item{transform}{can be either "l", "e1", or "e2".}

\item{k}{scaling parameter.}

\item{normalise}{normalise distances to one; can be either true of false.}

\item{max_d}{object of mode \code{numeric};}
}
\value{
\code{dist2sim(x)} returns an object of class \code{matrix}.
}
\description{
Converts any matrix of pairwise distances into pairwise similarities.
}
\examples{
tree <- ape::rtree(5)
dist <- phy2dist(tree)
dist2sim(dist, "l")


}
