% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-distance.R
\docType{class}
\name{distance-class}
\alias{distance-class}
\title{distance-class}
\description{
Container for class \code{distance}.
}
\section{Fields}{

\describe{
\item{\code{phylo}}{object of class \code{phylo}.}

\item{\code{depth}}{object of class \code{numeric}.}
}}

