% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-distance.R, R/distance.R
\docType{class}
\name{distance-class}
\alias{distance-class}
\alias{show,distance-method}
\title{distance-class}
\usage{
\S4method{show}{distance}(object)
}
\arguments{
\item{object}{object of class \code{distance}}
}
\description{
Container for class \code{distance}.
}
\section{Fields}{

\describe{
\item{\code{distance}}{object of class \code{matrix}.}

\item{\code{datID}}{object of class \code{character}.}

\item{\code{components}}{object of class \code{list}.}
}}

