% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\alias{distance,matrix,character-method}
\alias{distance,matrix,missing-method}
\title{Generate distance object}
\usage{
distance(distance, dat_id)

\S4method{distance}{matrix,character}(distance, dat_id)

\S4method{distance}{matrix,missing}(distance, dat_id)
}
\arguments{
\item{distance}{distance matrix}

\item{dat_id}{object of class \code{character} denoting the type of diversity
being calculated. This can be "naive", "genetic", "taxonomic", and so on}
}
\value{
\code{distance()} returns an object of class \code{distance}.
}
\description{
Container for class \code{distance}.
}
