% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Generate distance object}
\usage{
distance(distance, datID, components)
}
\arguments{
\item{distance}{distance matrix}

\item{datID}{}

\item{components}{sdf}
}
\value{
\code{distance()} returns an object of class \code{distance}.
}
\description{
Generate distance object
}
