% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Generate distance object}
\usage{
distance(distance, what)
}
\arguments{
\item{depth}{object of class \code{numeric}}
}
\value{
distance() returns an object of class \code{distance}
}
\description{
distance
}
