% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.diversity.R
\name{diversity}
\alias{diversity}
\title{Calculate Diversity}
\usage{
diversity(measure, populations, qs, zmatrix = diag(nrow(pmatrix)))
}
\arguments{
\item{measure}{Diversity measure}

\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}
}
\value{
A list of length = length(measure), where each list item contains 
the diversity output for the corresponding measure
}
\description{
Calculates the diversity, subcommunity or supercommunity, of a series of 
columns representing independent subcommunity counts, for a series of 
orders, repesented as a vector of qs.
}

