% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen2dist.r
\name{gen2dist}
\alias{gen2dist}
\title{Genetic distance matrix}
\usage{
gen2dist(vcf)
}
\arguments{
\item{vcf}{object of class \code{vcfR}.}
}
\value{
\code{gen2dist(x)} returns an object of class \code{distance}
containing a \code{matrix} of pairwise genetic distances.
}
\description{
Converts a vcfR object to a matrix of pairwise genetic distances.
}
