% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacommunity.R
\docType{methods}
\name{metacommunity}
\alias{metacommunity}
\alias{metacommunity,data.frame,missing-method}
\alias{metacommunity,data.frame-method}
\alias{metacommunity,numeric,missing-method}
\alias{metacommunity,numeric-method}
\alias{metacommunity,matrix,missing-method}
\alias{metacommunity,matrix-method}
\alias{metacommunity,data.frame,matrix-method}
\alias{metacommunity,data.frame-method,matrix-method}
\alias{metacommunity,numeric,matrix-method}
\alias{metacommunity,numeric-method,matrix-method}
\alias{metacommunity,matrix,matrix-method}
\alias{metacommunity,missing,similarity-method}
\alias{metacommunity,similarity-method}
\alias{metacommunity,numeric,similarity-method}
\alias{metacommunity,data.frame,similarity-method}
\alias{metacommunity,matrix,similarity-method}
\title{Metacommunity}
\usage{
metacommunity(partition, similarity)

\S4method{metacommunity}{data.frame,missing}(partition)

\S4method{metacommunity}{numeric,missing}(partition)

\S4method{metacommunity}{matrix,missing}(partition)

\S4method{metacommunity}{data.frame,matrix}(partition, similarity)

\S4method{metacommunity}{numeric,matrix}(partition, similarity)

\S4method{metacommunity}{matrix,matrix}(partition, similarity)

\S4method{metacommunity}{missing,similarity}(partition, similarity)

\S4method{metacommunity}{numeric,similarity}(partition, similarity)

\S4method{metacommunity}{data.frame,similarity}(partition, similarity)

\S4method{metacommunity}{matrix,similarity}(partition, similarity)
}
\arguments{
\item{partition}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types, columns as subcommunities, and elements containing  
the relative abundances of types in subcommunities. For phylogenetic 
diversity, see \emph{Details}.}

\item{similarity}{(optional) two-dimensional \code{matrix} of mode 
\code{numeric}, with rows as types, columns as types, and elements 
containing the pairwise similarity between types.}
}
\value{
\code{metacommunity()} returns an object of class 
\code{metacommunity} (see \emph{Fields}).
}
\description{
Functions to generate a \code{metacommunity} object.
}
\details{
When calculating phylogenetic diversity either:
\itemize{
\item set \code{partition} as the relative abundance of terminal taxa,  
in which case \code{similarity} (the pairwise similarity of historical species) is calculated from \code{phy2branch()} and 
\code{dist2sim()}; or
\item set \code{partition} as the relative abundance of historical species,
with \code{similarity} (the pairwise similarity of historical species) is
calculated from \code{phy2branch()} and 
\code{dist2sim()}.
}
}
\section{Fields}{

\describe{
\item{\code{type_abundance}}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types, columns as subcommunities, and elements containing  
relative abundances of types in subcommunities. In the phylogenetic case, 
this corresponds to the proportional abundance of historic species, which
is calculated from the proportional abundance of present day species.}

\item{\code{similarity}}{two-dimensional \code{matrix} of mode \code{numeric} with 
rows as types, columns as types, and elements containing pairwise 
similarities between types}

\item{\code{ordinariness}}{two-dimensional \code{matrix} of mode \code{numeric} 
with rows as types, columns as subcommunities, and elements containing the
ordinariness of types within subcommunities}

\item{\code{subcommunity_weights}}{\code{vector} of mode \code{numeric}; contains
subcommunity weights}

\item{\code{type_weights}}{two-dimensional \code{matrix} of mode \code{numeric}, 
with rows as types, columns as subcommunities, and elements containing 
weights of types within a subcommunity}

\item{\code{raw_abundance}}{[Phylogenetic] two-dimensional \code{matrix} of mode 
\code{numeric} with rows as types, columns as subcommunities, and elements 
containing the relative abundance of present day species}

\item{\code{raw_structure}}{[Phylogenetic] two-dimensional \code{matrix} of mode 
\code{numeric} with rows as historical species, columns as present day 
species, and elements containing historical species lengths within lineages}

\item{\code{parameters}}{[Phylogenetic] \code{tibble} containing parameters 
associated with each historic species in the phylogeny}
}}

\examples{
tree <- ape::rtree(n = 5)
tree$tip.label <- paste0("sp", seq_along(tree$tip.label))
partition <- cbind(a = c(1,1,1,0,0), b = c(0,1,0,1,1))
row.names(partition) <- tree$tip.label
partition <- partition / sum(partition)

similarity <- phy2branch(tree)
a <- metacommunity(partition, similarity)
b <- metacommunity(partition)

}
\seealso{
\code{\link{metacommunity-class}}
}
