% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadiv.R
\docType{methods}
\name{metadiv}
\alias{metadiv}
\alias{metadiv,powermean-method}
\alias{metadiv,relativeentropy-method}
\alias{metadiv,metacommunity-method}
\title{Calculate metacommunity-level diversity}
\usage{
metadiv(data, qs)

\S4method{metadiv}{powermean}(data, qs)

\S4method{metadiv}{relativeentropy}(data, qs)

\S4method{metadiv}{metacommunity}(data, qs)
}
\arguments{
\item{data}{\code{matrix} of mode \code{numeric}; containing diversity
components.}

\item{qs}{\code{vector} of mode \code{numeric}; parameter of conservatism.}
}
\value{
Returns a standard output of class \code{tibble}, with columns:
\itemize{
\item\code{measure}: raw or normalised, alpha, beta, rho, or gamma
\item\code{q}: parameter of conservatism
\item\code{type_level}: "metacommunity"
\item\code{type_name}: label attributed to type
\item\code{partition_level}: level of diversity, \emph{i.e.} metacommunity
\item\code{partition_name}: label attributed to partition
\item\code{diversity}: calculated metacommunity diversity
}
}
\description{
Generic function for calculating metacommunity-level diversity.
}
\details{
\code{data} may be input as three different classes:
\itemize{
\item{\code{powermean} calculates raw or normalised metacomunity alpha, rho
or gamma diversity by taking the powermean of diversity components}
\item{\code{relativeentropy} calculates raw or normalised metacommunity beta
diversity by taking the relative entropy of diversity components}
\item{\code{metacommunity} calculates all metacommunity measures of diversity}
}
}
\examples{
# Define metacommunity
pop <- data.frame(a = c(1,3), b = c(1,1))
pop <- pop / sum(pop)
meta <- metacommunity(pop)

# Calculate metacommunity gamma diversity (takes the power mean)
g <- raw_gamma(meta)
metadiv(g, 0:2)

# Calculate metacommunity beta diversity (takes the relative entropy)
b <- raw_beta(meta)
metadiv(b, 0:2)

# Calculate all measures of metacommunity diversity
metadiv(meta, 0:2)

}
\references{
Reeve, R., T. Leinster, C. Cobbold, J. Thompson, N. Brummitt, 
S. Mitchell, and L. Matthews. 2016. How to partition diversity. 
arXiv 1404.6520v3:1–9.
}
