% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{norm_sub_alpha}
\alias{norm_sub_alpha}
\title{Normalised subcommunity alpha diversity}
\usage{
norm_sub_alpha(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a standard output of class \code{tibble}, with columns:
\code{measure}, (norm alpha),
\code{q} (parameter of conservatism), 
\code{type_level} (), 
\code{type_name} (label attributed to type), 
\code{partition_level} (level of diversity, \emph{i.e.} subcommunity), 
\code{partition_name} (label attributed to partition), and 
\code{diversity}.
}
\description{
Calculates similarity-sensitive normalised subcommunity alpha diversity 
(the diversity of subcommunity \emph{j} in isolation. This measure may be 
calculated for a series of orders, repesented as a vector of \code{qs}.
}
\examples{
pop <- data.frame(a = c(1,3), b = c(1,1))
row.names(pop) <- paste0("sp", 1:2)
pop <- pop/sum(pop)
meta <- metacommunity(pop)

# Calculate normalised subcommunity alpha diversity
norm_sub_alpha(meta, 0:2)

}
\references{
Reeve, R., T. Leinster, C. Cobbold, J. Thompson, N. Brummitt, 
S. Mitchell, and L. Matthews. 2014. How to partition diversity. 
arXiv 1404.6520:1–9.
}
