% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{normalised.subcommunity.rho}
\alias{normalised.subcommunity.rho}
\title{Similarity-sensitive normalised subcommunity rho diversity}
\usage{
normalised.subcommunity.rho(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{normalised.subcommunity.rho()} is used to calculate the 
representativeness of subcommunity \emph{j}.
}
\details{
\code{normalised.subcommunity.rho()}, the inverse of \code{subcommunity.beta}), 
calculates the normalised subcommunity rho diversity of a series of columns 
representing independent subcommunities counts relative to a total 
supercommunity (by default the sum of the subcommunities), for a series of 
orders, repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate normalised subcommunity rho diversity
r <- normalised.rho(super)
subdiv(r, 0:2)
normalised.subcommunity.rho(super, 0:2)

}

