% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2branch.R
\name{phy2branch}
\alias{phy2branch}
\title{phy2branch}
\usage{
phy2branch(tree, depth = 1)
}
\arguments{
\item{tree}{object of class \code{phylo}.}

\item{depth}{proportion of total tree height to be conserved (taken as
a proportion from the heighest tip). Describes how much evolutionary history 
should be retained, with 0 marking the date of the most recent tip, and 1 
(the default) marking the most recent common ancestor. Numbers greater than 
1 extend the root of the tree.}
}
\value{
\code{phy2sim()} returns an object of class \code{similarity}.
}
\description{
Packages phylo object and depth into an object of class \code{similarity}.
}
