% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2dist.R
\name{phy2dist}
\alias{phy2dist}
\title{Phylogenetic pairwise tip distance matrix}
\usage{
phy2dist(tree, precompute_dist = TRUE)
}
\arguments{
\item{tree}{object of class \code{phylo}.}

\item{precompute_dist}{object of class \code{logical} or \code{numeric}.
When TRUE (by default) a distance matrix is generated and stored in slot
\code{distance}, when FALSE no distance matrix is generated, and when numeric
a distance matrix is generated until the number of species exceeds the
defined value.}
}
\value{
\code{phy2sim(x)} returns an object of class \code{distance}
containing a \code{matrix} of pairwise tip-to-tip distances.
}
\description{
Converts any \code{phylo} object to a matrix of pairwise tip-to-tip distances.
}
