% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_abundance.R
\name{phy_abundance}
\alias{phy_abundance}
\title{phy_abundance}
\usage{
phy_abundance(partition, ps)
}
\arguments{
\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with 
rows as types, columns as subcommunities, and elements containing relative 
abundances of types in subcommunities. In the case of phylogenetic 
metacommunities, these are the relative abundances of terminal taxa.}

\item{ps}{\code{phy_struct()} output.}
}
\description{
Calculates the relative abundance of historic species
}
\examples{


}
