% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_abundance.R
\name{phy_abundance}
\alias{phy_abundance}
\title{Relative abundance of historical species}
\usage{
phy_abundance(partition, structure_matrix)
}
\arguments{
\item{partition}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types (species), columns as subcommunities, and each
element containing the relative abundance of types in each subcommunity
relative to the metacommunity as a whole. In the phylogenetic case, this
corresponds to the proportional abundance of historical species, which is
calculated from the proportional abundance of terminal taxa}

\item{structure_matrix}{\code{output$structure} of \code{phy_struct()}.}
}
\description{
Calculates the relative abundance of historical species.
}
