% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-phy_struct.R
\docType{class}
\name{phy_struct-class}
\alias{phy_struct-class}
\title{phy_struct-class}
\value{
An object of class \code{phy_struct}. Builds on class \code{phylo} and 
therefore contains the following components:
\tabular{ll}{
\code{edge} \tab two-column \code{matrix} of mode numeric; each row  
represents an edge of the tree. Nodes and tips are symbolized with numbers.
Tips are numbered 1, 2, ...,N and inner nodes are numbered N+1, N+2, ... 
For each row, the first column gives the ancestor. \cr
\code{tip.label} \tab \code{vector} of mode character; giving the names of the 
tips ordered by the (positive) number in edge. \cr
\code{edge.length} \tab (optional) \code{vector} of mode numeric; giving the 
lengths of the branches given by edge. \cr
\code{Nnode} \tab \code{integer}; number of internal nodes. \cr
}
}
\description{
Container for various parameters associated with calculating phylogenetic 
diversity.
}
\section{Fields}{

\describe{
\item{\code{structure}}{\code{matrix} object}

\item{\code{parameters}}{object of class \code{tbl_df}}
}}

