% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_struct.R
\docType{methods}
\name{phy_struct}
\alias{phy_struct}
\alias{is.phy_struct}
\alias{show,phy_struct-method}
\title{Convert to phy_struct}
\usage{
phy_struct(tree, partition)

is.phy_struct(x)

\S4method{show}{phy_struct}(object)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with 
rows as types, columns as subcommunities, and elements containing relative 
abundances of types in subcommunities. In the case of phylogenetic 
metacommunities, these are the relative abundances of terminal taxa.}

\item{x}{any R object}

\item{object}{object of class \code{phy_struct}}
}
\value{
Returns an object of class \code{phy_struct}, which contains: 

\tabular{ll}{
\code{/@structure} \tab - each row denotes historic species, columns denote 
terminal taxa, and elements contain branch lengths \cr
\code{/@parameters} \tab - information associated with each historic species \cr
}

Additional slots (accessed with @, not $) include:

\tabular{ll}{
\code{historic.species} \tab - a data_frame containing components associated 
with historic species \cr
\code{terminal_taxa} \tab - na data_frame containing components associated 
with terminal taxa \cr
\code{Tbar} \tab - \cr
\code{structure} \tab -  \cr
}
}
\description{
Converts an object into class \code{phylo} into class \code{phy_struct}.
}
\examples{
tree <- ape::rtree(n = 5)
tree$tip.label <- paste0("sp", seq_along(tree$tip.label))

partition <- cbind(a = c(0,1,1,1,0), b = c(1,0,0,0,1))
row.names(partition) <- tree$tip.label 
partition <- partition / sum(partition)

res <- phy_struct(tree, partition)

}
