% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powermean.R
\docType{methods}
\name{powermean}
\alias{powermean}
\alias{is.powermean}
\alias{show,powermean-method}
\title{Coerce to powermean}
\usage{
powermean(results, meta, tag)

is.powermean(x)

\S4method{show}{powermean}(object)
}
\arguments{
\item{results}{two-dimensional \code{matrix} of mode \code{numeric};
contains values calculated from diversity-term functions, e.g.
\code{norm_alpha()}, \code{raw_beta}, etc.}

\item{meta}{object of class \code{metacommunity}; contains proportional
abundance of types, pair-wise similarity, and other associated variables.}

\item{tag}{object of class \code{character}; contains an identifier
associated with the calculated diversity-term, e.g.
"norm_alpha", "raw_beta", etc.}

\item{x}{any R object}

\item{object}{object of class \code{powermean}}
}
\value{
\code{powermean(x)} returns an object of class \code{powermean}.

\code{is.powermean(x)} returns TRUE if its argument is a
powermean, FALSE otherwise.

\code{print(x)} prints an object object of class \code{powermean}
}
\description{
Functions to check if an object is a \code{powermean} or coerce an
object into a \code{powermean}.
}
\details{
\code{powermean} is a class of object output from calculating
\code{raw_alpha}, \code{norm_alpha}, \code{raw_rho},
\code{norm_rho}, and \code{raw_gamma} diversity terms;
\code{raw_beta} and \code{norm_beta} are output as
\code{relativeentropy} objects.

To calculate the subcommunity or metacommunity diversity, an object of
class \code{powermean} is input to \code{subdiv()} or \code{metadiv()},
respectively.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate subcommunity raw alpha diversity (takes the powermean)
a <- raw_alpha(meta)
class(a)

}
