% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.powermean.R
\docType{methods}
\name{powermean}
\alias{as.powermean}
\alias{is.powermean}
\alias{powermean}
\alias{show,powermean-method}
\title{Coerse to powermean}
\usage{
powermean(results, super, tag)

as.powermean(results, super, tag)

is.powermean(x)

\S4method{show}{powermean}(object)
}
\arguments{
\item{results}{two-dimensional \code{matrix} of mode \code{numeric};}

\item{super}{object of class \code{supercommunity};}

\item{tag}{object of class \code{character};}

\item{x}{any R object}

\item{object}{object of class \code{powermean}}
}
\value{
object of class \code{powermean}

\code{is.powermean(x)} returns TRUE if its argument is a 
powermean, FALSE otherwise.
}
\description{
Functions to check if an object is a \code{powermean} or coerce an  
object into a \code{powermean}.
}
\details{

}
\examples{

}

