% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qD.R
\name{qD}
\alias{qD}
\title{Hill number / naive diversity with no similarity}
\usage{
qD(populations, qs)
}
\arguments{
\item{populations}{- population counts or proportions}

\item{qs}{- vector of values of parameter q}
}
\value{
data frame of diversities, columns representing populations, and 
rows representing values of q
}
\description{
Calculates the diversity of a series of columns representing independent
populations, for a series of orders, repesented as a vector of qs.
}
\details{
\code{qD} is used to calculate the diversity of a population
(in the naive-type case).
}
\examples{
#

}

