% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qD.single.R
\name{qD.single}
\alias{qD.single}
\title{Hill number / naive diversity with no similarity of a single population}
\usage{
qD.single(proportions, q)
}
\arguments{
\item{proportions}{Relative proportions of different individuals / types 
in population}

\item{q}{Order of diversity measurement}
}
\value{
as
}
\description{
Calculates the Hill number / naive diversity of order q of a population
with given relative proportions
}
\details{
asd
}
\examples{

}

