% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qDZ.R
\name{qDZ}
\alias{qDZ}
\title{Similarity-sensitive diversity}
\usage{
qDZ(populations, qs)
}
\arguments{
\item{populations}{object of class \code{initDiv}}

\item{qs}{Vector of values of parameter q}
}
\value{
Data frame of diversities, columns representing populations, and 
rows representing values of q
}
\description{
Calculates the diversity of a series of columns representing independent
populations, for a series of orders, repesented as a vector of qs.
}
\details{

}
\examples{

}

