% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\name{qDZ}
\alias{qDZ}
\title{Similarity-sensitive diversity}
\usage{
qDZ(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}
}
\value{
Data frame of diversities, columns representing populations, and 
rows representing values of q
}
\description{
Calculates the diversity of a series of columns representing independent
populations, for a series of orders, repesented as a vector of qs.
}

