% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{raw.subcommunity.rho}
\alias{raw.subcommunity.rho}
\title{Similarity-sensitive raw subcommunity rho diversity}
\usage{
raw.subcommunity.rho(meta, qs)
}
\arguments{
\item{meta}{object of class \code{metacommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:  
\code{partition} (label attributed to partition), \code{q} (parameter of 
conservatism), \code{diversity}, \code{community} (level of diversity, 
\emph{i.e.} subcommunity, community, or metacommunity), and \code{measure} 
(alpha, beta, rho, or gamma).
}
\description{
\code{raw.subcommunity.rho()} is used to calculate the redundancy of 
subcommunity \emph{j}.
}
\details{
\code{raw.subcommunity.rho()}, the inverse of \code{raw.subcommunity.beta()}, 
calculates the raw subcommunity rho diversity of a series of columns  
representing independent subcommunity counts relative to the
metacommunity as a whole (by default the sum of the subcommunities). This 
measure may be calculated for a series of orders, repesented as a vector 
of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate raw subcommunity rho diversity
r <- raw.rho(meta)
subdiv(r, 0:2)
raw.subcommunity.rho(meta, 0:2)

}

