% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-components.R
\name{raw_gamma}
\alias{raw_gamma}
\title{Low level diversity components: gamma}
\usage{
raw_gamma(meta)
}
\arguments{
\item{meta}{object of class \code{metacommunity}.}
}
\value{
Returns an object of class \code{powermean}.
}
\description{
Calculates the low-level diversity component necessary for calculating gamma
diversity.
}
\details{
Values generated from \code{raw_gamma()} may be input into \code{subdiv()} and
\code{metadiv()} to calculate subcommunity/metacommunity gamma diversity.
}
\examples{
pop <- sample(1:50, 5)
meta <- metacommunity(pop)

# Calculate gamma component
a <- raw_gamma(meta)

}
