% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-rdiv.R, R/rdiv.R
\docType{class}
\name{rdiv-class}
\alias{rdiv-class}
\alias{show,rdiv-method}
\title{rdiv-class}
\usage{
\S4method{show}{rdiv}(object)
}
\arguments{
\item{object}{object of class \code{rdiv}}
}
\description{
Container for class \code{rdiv}.
}
