% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-rdiversity.R
\docType{class}
\name{rdiv-class}
\alias{rdiv}
\alias{rdiv-class}
\alias{show,rdiv-method}
\title{Class 'rdiv'}
\usage{
\S4method{show}{rdiv}(object)
}
\arguments{
\item{object}{object of class \code{rdiv}}
}
\description{
Define S4 class \code{rdiv}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Prints diversity results
}}
\section{Fields}{

\describe{
\item{\code{measure}}{object of class \code{character}; e.g. 'raw_subcommunity_alpha'}

\item{\code{tag}}{object of class \code{formula}; e.g. bquote('Subcommunity' ~ bar(alpha))}

\item{\code{level}}{object of class \code{character}; e.g. 'subcommunity'}
}}

