% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdiversity-package.R
\docType{package}
\name{rdiversity-package}
\alias{rdiversity-package}
\alias{rdiversity}
\title{rdiversity}
\description{
\code{rdiversity} is a package for R based around a framework of 
similarity-sensitive diversity measures. It provides functionality for 
measuring alpha, beta and gamma diversity of metacommunities (e.g. 
ecosystems) and their constituent subcommunities, where similarity may be 
defined as taxonomic, phenotypic, genetic, phylogenetic, functional, and so 
on. It uses the diversity measures described in the arXiv paper 
arXiv:1404.6520 (q-bio.QM), How to partition diversity.
}
\details{
For more information go to our gitHub page 
(https://github.com/boydorr/rdiversity).
}
