\name{RDiversity-package}
\alias{RDiversity-package}
\alias{RDiversity}
\docType{package}
\title{
\packageTitle{RDiversity}
}
\description{
\packageDescription{RDiversity}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RDiversity}
\packageIndices{RDiversity}
~~ An overview of how to use the ~~
~~ package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{RDiversity}

Maintainer: \packageMaintainer{RDiversity}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
