% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-rdphylo.R
\docType{class}
\name{rdphylo-class}
\alias{rdphylo-class}
\title{Class 'rdphylo'}
\description{
Class 'rdphylo'
}
\section{Fields}{

\describe{
\item{\code{hs.name}}{\code{vector} of length \emph{hS}; historic species names}

\item{\code{hs.pds}}{\code{vector} of length \emph{hS}; descendant present day 
species}

\item{\code{hs.edge}}{\code{matrix}; ancestral and descendant nodes}

\item{\code{hs.abundance}}{\code{vector} of length \emph{hS}; historic species 
abundance}

\item{\code{Lj}}{\code{vector} of length \emph{S}; total evolutionary change of 
present day species}

\item{\code{Tbar}}{\code{numeric} element; mean total evolutionary change over 
present day species}
}}

