% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-rdphylo.R
\docType{class}
\name{RDphylo-class}
\alias{RDphylo}
\alias{RDphylo-class}
\alias{show,RDphylo-method}
\title{Class 'RDphylo'}
\usage{
\S4method{show}{RDphylo}(object)
}
\arguments{
\item{object}{object of class \linkS4class{RDphylo}}
}
\value{
An object of class \code{RDphylo}. Builds on class \code{phylo} and 
therefore contains the following components:
\tabular{ll}{
\code{edge} \tab two-column \code{matrix} of mode numeric; each row  
represents an edge of the tree. Nodes and tips are symbolized with numbers.
Tips are numbered 1, 2, ...,N and inner nodes are numbered N+1, N+2, ... 
For each row, the first column gives the ancestor. \cr
\code{tip.label} \tab \code{vector} of mode character; giving the names of the 
tips ordered by the (positive) number in edge. \cr
\code{edge.length} \tab (optional) \code{vector} of mode numeric; giving the 
lengths of the branches given by edge. \cr
\code{Nnode} \tab \code{integer}; number of internal nodes. \cr
}
}
\description{
Container for various parameters associated with calculating phylogenetic 
diversity.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: prints various parameters associated with the phylogeny
}}
\section{Fields}{

\describe{
\item{\code{hs.edge}}{two-column \code{matrix} of mode numeric; each row represents
a historic species. Nodes and tips are symbolized with numbers. Tips are 
numbered 1, 2, ..., and nodes are numbered after the tips. 
For each row, the first column gives the ancestor.}

\item{\code{hs.name}}{\code{vector} of more character; giving the names of the 
historic species, ordered by the number in hs.edge.}

\item{\code{hs.pds}}{\code{vector} of mode numeric; giving the present day 
species descendant from each historic species, ordered by the number in 
hs.edge.}

\item{\code{hs.abundance}}{\code{vector} of mode numeric; giving the relative 
abundance of historic species, ordered by the number in hs.edge.}

\item{\code{Lj}}{\code{vector} of length \emph{S}; giving the total evolutionary 
change of present day species, ordered by the number in edge.}

\item{\code{Tbar}}{\code{numeric} element; giving the mean total evolutionary 
change over present day species.}
}}

