% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rdphylo.R
\docType{methods}
\name{rdphylo}
\alias{as.rdphylo}
\alias{is.rdphylo}
\alias{rdphylo}
\alias{show,rdphylo-method}
\title{Convert to rdphylo}
\usage{
rdphylo(pds.abundance, tree)

as.rdphylo(pds.abundance, tree)

is.rdphylo(x)

\S4method{show}{rdphylo}(object)
}
\arguments{
\item{pds.abundance}{\code{vector} of length \eqn{S}; proportional
abundance of present-day species}

\item{tree}{object of class \code{phylo}}

\item{x}{any R object}

\item{object}{object of class \code{rdphylo}}
}
\value{
Returns an object of class \code{rdphylo}, which extends class
\code{phylo} and therefore inherits components associated
with terminal taxa (see \code{\link{phylo}}):

\tabular{ll}{
\code{edge} \tab - each row denotes the ancestral and descendant nodes of
an edge (branch) of the tree \cr
\code{edge.length} \tab - (optional) length of each branch of the tree \cr
\code{tip.label} \tab - labels associated with present-day species
(terminal taxa)  \cr
\code{Nnode} \tab - number of internal nodes \cr
\code{node.label} \tab - (optional) labels associated with each node in
the tree \cr
\code{root.edge} \tab - (optional) length of the branch at the root \cr
}

Additional slots (accessed with @, not $) include:

\tabular{ll}{
\code{historic.species} \tab - a data_frame containing components associated 
with historic species \cr
\code{terminal.taxa} \tab - na data_frame containing components associated 
with terminal taxa \cr
\code{Tbar} \tab - \cr
\code{structure} \tab -  \cr
}
}
\description{
Converts an object into class \code{phylo} into class \code{rdphylo}.
}
\examples{
tree <- ape::rtree(n = 5)
pop <- rep(1/5, 5)
rdtree <- rdphylo(pop, tree)

pds.abundance <- data.frame(a = c(0,1,1,1,0), b = c(1,0,0,0,1))
pds.abundance <- check_partition(pds.abundance)

# Extract components of historic species
rdtree@historic.species

}

