% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdplot.R
\name{rdplot}
\alias{rdplot}
\title{rdplot()}
\usage{
rdplot(data, plot.title)
}
\arguments{
\item{data}{output}

\item{plot.title}{object of class \code{character}}
}
\description{
Simple function to plot the diversity profile of a single diversity measure
}
