% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-relativeentropy.R
\docType{class}
\name{relativeentropy-class}
\alias{relativeentropy-class}
\title{relativeentropy-class}
\description{
Container for class \code{relativeentropy}.
}
\section{Fields}{

\describe{
\item{\code{results}}{object of class \code{matrix} of mode \code{numeric}; contains
diversity term values output from \code{raw_beta()} or \code{norm_beta()}}

\item{\code{measure}}{measure}

\item{\code{type_abundance}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains proportional abundance of \emph{types} in the subcommunity 
as a fraction of the metacommunity as a whole (in the phylogenetic case, 
this corresponds to the proportional abundance of historic species, which
is calculated from the proportional abundance of present day species)}

\item{\code{ordinariness}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains ordinariness of types}

\item{\code{subcommunity_weights}}{\code{vector} of mode \code{numeric}; contains
subcommunity weights}

\item{\code{type_weights}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains weight of types within a subcommunity}
}}

\references{
Reeve, R., T. Leinster, C. Cobbold, J. Thompson, N. Brummitt, 
S. Mitchell, and L. Matthews. 2014. How to partition diversity. 
arXiv 1404.6520:1–9.
}
