% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-similarity.R, R/similarity.R
\docType{class}
\name{similarity-class}
\alias{similarity-class}
\alias{show,similarity-method}
\title{similarity-class}
\usage{
\S4method{show}{similarity}(object)
}
\arguments{
\item{object}{object of class \code{similarity}}
}
\description{
Container for class \code{similarity}.
}
\section{Fields}{

\describe{
\item{\code{similarity}}{object of class \code{matrix}.}

\item{\code{datID}}{object of class \code{character}.}

\item{\code{components}}{object of class \code{list}.}

\item{\code{parameters}}{object of class \code{list}.}
}}

