% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-similarity.R, R/similarity.R
\docType{class}
\name{similarity-class}
\alias{similarity-class}
\alias{show,similarity-method}
\title{similarity-class}
\usage{
\S4method{show}{similarity}(object)
}
\arguments{
\item{object}{object of class \code{similarity}}
}
\description{
Container for class \code{similarity}.
}
\section{Fields}{

\describe{
\item{\code{similarity}}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types, columns as types, and elements containing the pairwise
similarity of types}

\item{\code{dat_id}}{object of class \code{character} describing the class of
distance / similarity being used, e.g. "naive", "taxonomic", and so on}

\item{\code{components}}{list containing the components necessary to calculate
similarity. This list is empty when \code{precompute_dist = TRUE} when
calculating distance. When a pairwise distance matrix is too large and
\code{precompute_dist = FALSE}, this list contains all the information
required to calculate pairwise distance between types}

\item{\code{parameters}}{list containing parameters associated with
converting pairwise distances to similarities (the \code{dist2sim()}
arguments)}
}}

