% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\alias{similarity,matrix,character-method}
\alias{similarity,matrix,missing-method}
\title{Generate similarity object}
\usage{
similarity(similarity, dat_id)

\S4method{similarity}{matrix,character}(similarity, dat_id)

\S4method{similarity}{matrix,missing}(similarity, dat_id)
}
\arguments{
\item{similarity}{similarity matrix}

\item{dat_id}{object of class \code{character} denoting the type of diversity
being calculated. This can be "naive", "genetic", "taxonomic", and so on}
}
\value{
\code{similarity()} returns an object of class \code{similarity}.
}
\description{
Container for class \code{similarity}.
}
