% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Calculate similarity}
\usage{
similarity(phylo, depth = 1)
}
\arguments{
\item{phylo}{object of class \code{phylo}}

\item{depth}{object of class \code{numeric}}
}
\value{
object of class \code{similarity}
}
\description{
Functions to check if an object is a \code{similarity}, or coerse an  
object into a \code{similarity}; output of \code{phy2sim}.
}
