% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_phylo.R
\name{similarity_phylo}
\alias{similarity_phylo}
\title{Phylogenetic Similarity Matrix}
\usage{
similarity_phylo(tree, pds.abundance = rep(1/length(tree$tip.label),
  length(tree$tip.label)))
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{pds.abundance}{vector of length {S}; relative abundance of present-day 
species}
}
\value{
\eqn{hS x hS} matrix; pair-wise similarity of historic species
}
\description{
Function to calculate phylogenetic similarity matrix from a phylogeny.
}
\details{

}
\examples{

tree <- ape::rtree(n = 5)
Z <- similarity_phylo(tree)


}

