% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smatrix.R
\name{smatrix}
\alias{smatrix}
\title{Phylogenetic similarity matrix (ultrametric)}
\usage{
smatrix(ps)
}
\arguments{
\item{ps}{\code{phy_struct()} output.}
}
\value{
Returns an \eqn{hS x hS} matrix; pair-wise ultrametric-similarity of
historic species.
}
\description{
Function to calculate an ultrametric-similarity matrix.
}
\examples{
tree <- ape::rtree(10)
partition <- matrix(rep(1, 500), nrow = 10)
partition <- as.data.frame(partition / sum(partition))
rownames(partition) <- tree$tip.label
colnames(partition) <- LETTERS[seq_len(ncol(partition))]
ps <- phy_struct(tree, partition)

}
