% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smatrix.R
\name{smatrix}
\alias{smatrix}
\title{Phylogenetic similarity matrix (ultrametric)}
\usage{
smatrix(ps)
}
\arguments{
\item{ps}{\code{phy_struct()} output.}
}
\value{
Returns an \eqn{hS x hS} matrix; pair-wise ultrametric-similarity of
historic species.
}
\description{
Function to calculate an ultrametric-similarity matrix.
}
