% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\docType{data}
\name{subcommunity.alpha.bar}
\alias{subcommunity.alpha.bar}
\title{Similarity-sensitive Normalised subcommunity.alpha}
\format{An object of class \code{diversity} of length 1.}
\usage{
subcommunity.alpha.bar
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}

\item{normalise}{Normalise probability distribution to sum to 1}
}
\value{
An array of diversities, first dimension representing 
sub-communities, and last representing values of q
}
\description{
Calculates the diversity of a series of columns representing
independent subcommunity counts, for a series of orders, repesented as
a vector of qs
}
\keyword{datasets}

