% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.alpha.bar.R
\name{subcommunity.alpha.bar}
\alias{subcommunity.alpha.bar}
\title{Similarity-sensitive Normalised subcommunity.alpha}
\usage{
subcommunity.alpha.bar(populations, qs)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
An array of diversities, first dimension representing 
sub-communities, and last representing values of \emph{q}
}
\description{
\code{subcommunity.alpha} calculates the diversity of a series of columns 
representing independent subcommunity counts, for a series of orders 
repesented as a vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace=T),
                        subcommunityB = sample(1:50, 5, replace=T))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

# Calculate diversity
subcommunity.alpha.bar(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

