% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.alpha.bar.R
\name{subcommunity.alpha.bar}
\alias{subcommunity.alpha.bar}
\title{Similarity-sensitive Normalised subcommunity.alpha}
\usage{
subcommunity.alpha.bar(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}
}
\value{
An array of diversities, first dimension representing 
sub-communities, and last representing values of \emph{q}
}
\description{
\code{subcommunity.alpha} calculates the diversity of a series of columns 
representing independent subcommunity counts, for a series of orders 
repesented as a vector of \code{qs}.
}
\seealso{
\code{\link{subcommunity.alpha}}, \code{\link{supercommunity.A}}, \code{\link{supercommunity.A.bar}}
}

