% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{subcommunity.alpha.bar}
\alias{subcommunity.alpha.bar}
\title{Similarity-sensitive normalised subcommunity.alpha}
\usage{
subcommunity.alpha.bar(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{subcommunity.alpha.bar()} is used to calculate the 
similarity-sensitive diversity of subcommunity \emph{j} in isolation.
}
\details{
\code{subcommunity.alpha.bar()} calculates the normalised subcommunity 
alpha diversity of a series of columns representing independent subcommunity 
counts relative to the supercommunity as a whole (by default the sum of the 
subcommunities). This measure may be calculated for a series of orders, 
repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate normalised subcommunity alpha diversity (takes the power mean)
a <- alphabar(super)
subdiv(a, 0:2)

}

