% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.alpha.bar.R
\name{subcommunity.alpha.bar}
\alias{subcommunity.alpha.bar}
\title{Similarity-sensitive Normalised subcommunity.alpha}
\usage{
subcommunity.alpha.bar(populations, qs)
}
\arguments{
\item{populations}{object of class \code{collection}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
An array of diversities, first dimension representing 
sub-communities, and last representing values of \emph{q}
}
\description{
\code{subcommunity.alpha} calculates the diversity of a series of columns 
representing independent subcommunity counts, for a series of orders 
repesented as a vector of \code{qs}.
}
\details{
\code{subcommunity.alpha.bar()} is used to calculate the 
similarity-sensitive diversity of subcommunity \emph{j} in isolation.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Coerse object into a supercommunity
data <- as.supercommunity(population)

# Calculate diversity
subcommunity.alpha.bar(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

