% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.alpha.R
\name{subcommunity.alpha}
\alias{subcommunity.alpha}
\title{Similarity-sensitive Raw subcommunity.alpha}
\usage{
subcommunity.alpha(populations, qs, Z = diag(nrow(populations)),
  normalise = F)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
Returns an array of diversities, first dimension representing 
subcommunities, and last representing values of \emph{q}
}
\description{
Calculates the diversity of a series of columns representing
independent subcommunity counts, for a series of orders, repesented as
a vector of \code{qs}.
}
\seealso{
\code{\link{subcommunity.alpha.bar}}, \code{\link{supercommunity.A}}, \code{\link{supercommunity.A.bar}}
}

