% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.beta.bar.R
\name{subcommunity.beta.bar}
\alias{subcommunity.beta.bar}
\title{Similarity-sensitive Normalised subcommunity.beta diversity}
\usage{
subcommunity.beta.bar(populations, qs, Z = diag(nrow(populations)), ...)
}
\arguments{
\item{populations}{Population counts or proportions - single vector or matrix}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}

\item{...}{}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of q
}
\description{
Calculates the diversity of a series of columns representing independent
subcommunities counts relative to a total supercommunity (by default the 
sum of the sub-communities), for a series of orders, repesented as a 
vector of qs.
}
\seealso{
\code{\link{subcommunity.beta}}
}

