% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.beta.R
\name{subcommunity.beta}
\alias{subcommunity.beta}
\title{Similarity-sensitive Raw subcommunity.beta diversity}
\usage{
subcommunity.beta(populations, qs, Z = diag(nrow(populations)),
  normalise = F)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of \emph{q}
}
\description{
Calculates the diversity of a series of columns representing independent
subcommunities counts relative to a total supercommunity (by default the 
sum of the subcommunities), for a series of orders, repesented as a 
vector of \code{qs}.
}
\seealso{
\code{\link{subcommunity.beta.bar}}, \code{\link{supercommunity.B}}, \code{\link{supercommunity.B.bar}}
}

