% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.gamma.bar.R
\name{subcommunity.gamma.bar}
\alias{subcommunity.gamma.bar}
\title{Similarity-sensitive Normalised subcommunity.gamma diversity}
\usage{
subcommunity.gamma.bar(populations, qs, ...)
}
\arguments{
\item{populations}{object of class \code{initDiv}}

\item{qs}{\code{vector} of \emph{q} values}

\item{...}{additional arguments}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of \emph{q}
}
\description{
Calculates the diversity of a series of columns representing independent 
subcommunities counts relative to a total supercommunity (by default the 
sum of the subcommunities), for a series of orders, repesented as a  
vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

# Calculate diversity
subcommunity.gamma.bar(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

