% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.gamma.R
\name{subcommunity.gamma}
\alias{subcommunity.gamma}
\title{Similarity-sensitive Raw subcommunity.gamma diversity}
\usage{
subcommunity.gamma(populations, qs, normalise = F)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of \emph{q}
}
\description{
Calculates the diversity of a series of columns representing independent 
subcommunities counts relative to a total supercommunity (by default the 
sum of the subcommunities), for a series of orders, repesented as a  
vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace=T),
                        subcommunityB = sample(1:50, 5, replace=T))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

# Calculate diversity
subcommunity.gamma(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

