% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\docType{data}
\name{subcommunity.gamma}
\alias{subcommunity.gamma}
\title{Similarity-sensitive Raw subcommunity.gamma diversity}
\format{An object of class \code{diversity} of length 1.}
\usage{
subcommunity.gamma
}
\arguments{
\item{populations}{Population counts or proportions; single vector or matrix}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}

\item{normalise}{Normalise probability distribution to sum to 1}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of q
}
\description{
Calculates the diversity of a series of columns representing independent 
sub-communities counts relative to a total supercommunity (by default the 
sum of the sub-communities), for a series of orders, repesented as a  
vector of qs.
}
\keyword{datasets}

