% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.rho.R
\name{subcommunity.rho}
\alias{subcommunity.rho}
\title{Similarity-sensitive Normalised subcommunity.rho diversity}
\usage{
subcommunity.rho(populations, qs, Z = diag(nrow(populations)), ...)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of \emph{q}
}
\description{
The inverse of the similarity-sensitive Normalised subcommunity.beta 
diversity; Calculates the diversity of a series of columns representing 
independent subcommunities counts relative to a total supercommunity (by 
default the sum of the subcommunities), for a series of orders, repesented 
as a vector of \code{qs}.
}
\seealso{
\code{\link{subcommunity.rho.bar}}, \code{\link{supercommunity.R}}, \code{\link{supercommunity.R.bar}}
}

