% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{subcommunity.rho}
\alias{subcommunity.rho}
\title{Similarity-sensitive raw subcommunity.rho diversity}
\usage{
subcommunity.rho(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{subcommunity.rho()} is used to calculate the redundancy of 
subcommunity \emph{j}.
}
\details{
\code{subcommunity.rho()}, the inverse of \code{subcommunity.beta.bar}, 
calculates the subcommunity rho diversity of a series of columns  
representing independent subcommunity counts relative to the
supercommunity as a whole (by default the sum of the subcommunities). This 
measure may be calculated for a series of orders, repesented as a vector 
of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate subcommunity rho diversity (takes the power mean)
a <- rho(super)
subdiv(a, 0:2)

}

