% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcommunity.rho.R
\name{subcommunity.rho}
\alias{subcommunity.rho}
\title{Similarity-sensitive Normalised subcommunity.rho diversity}
\usage{
subcommunity.rho(populations, qs, ...)
}
\arguments{
\item{populations}{object of class \code{collection}}

\item{qs}{\code{vector} of \emph{q} values}

\item{...}{additional arguments}
}
\value{
Data frame of diversities, columns representing populations, and
rows representing values of \emph{q}
}
\description{
\code{subcommunity.rho()} is used to calculate the redundancy of 
subcommunity \emph{j}.
}
\details{
\code{subcommunity.rho()} (the inverse of \code{subcommunity.beta.bar}) 
calculates the subcommunity rho diversity of a series of columns  
representing independent subcommunities counts relative to a total 
supercommunity (by default the sum of the subcommunities), for a series of 
orders, repesented as a vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Coerse object into a supercommunity
data <- supercommunity(population)

# Calculate diversity
subcommunity.rho(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

