% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdiv.R
\docType{methods}
\name{subdiv}
\alias{subdiv}
\alias{subdiv,powermean-method}
\alias{subdiv,relativeentropy-method}
\alias{subdiv,supercommunity-method}
\title{Calculate subcommunity diversity}
\usage{
subdiv(data, qs)

\S4method{subdiv}{powermean}(data, qs)

\S4method{subdiv}{relativeentropy}(data, qs)

\S4method{subdiv}{supercommunity}(data, qs)
}
\arguments{
\item{data}{two-dimensional \code{matrix} of mode \code{numeric}; diversity 
components.}

\item{qs}{\code{vector} of mode \code{numeric}; parameter of conservatism.}
}
\value{
Returns a \code{data_frame} containing all of the diversity measures.


}
\description{
Generic function for calculating subcommunity diversity.
}
\details{
\code{data} may be input as three different classes:
\itemize{
\item{\code{powermean} calculates raw and normalised subcomunity alpha, rho 
or gamma diversity by taking the powermean of diversity components}
\item{\code{relativeentropy} calculates raw or normalised subcommunity beta 
diversity by taking the relative entropy of diversity components}
\item{\code{supercommunity} caculates all subcommunity measures of diversity}
}
}
\examples{
# Calculate the diversity of a single population
pop <- data.frame(a = c(1,3), b = c(1,1))
pop <- pop/sum(pop)
super <- supercommunity(pop)

# Subcommunity gamma diversity (takes the power mean)
g <- raw.gamma(super)
subdiv(g, 0:2)

# Subcommunity beta diversity (takes the relative entropy)
b <- raw.beta(super)
subdiv(b, 0:2)

# All measures of subcommunity diversity
subdiv(super, 0:2)

}

