% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdiv.R
\docType{methods}
\name{subdiv}
\alias{subdiv}
\alias{subdiv,powermean-method}
\alias{subdiv,relativeentropy-method}
\title{Calculate subcommunity diversity}
\usage{
subdiv(data, qs)

\S4method{subdiv}{powermean}(data, qs)

\S4method{subdiv}{relativeentropy}(data, qs)
}
\arguments{
\item{data}{two-dimensional \code{matrix} of mode \code{numeric}; diversity 
terms.}

\item{qs}{\code{vector} of mode \code{numeric}; parameter of conservatism.}
}
\value{
Returns a two-dimensional \code{matrix} of mode \code{numeric}.
}
\description{
Calculates subcommunity diversity by taking the powermean of diversity 
terms.
}
\details{
as ds
}
\examples{
as

}

