% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdiv.R
\docType{methods}
\name{subdiv}
\alias{subdiv}
\alias{subdiv,metacommunity-method}
\alias{subdiv,powermean-method}
\alias{subdiv,relativeentropy-method}
\title{Calculate subcommunity diversity}
\usage{
subdiv(data, qs)

\S4method{subdiv}{powermean}(data, qs)

\S4method{subdiv}{relativeentropy}(data, qs)

\S4method{subdiv}{metacommunity}(data, qs)
}
\arguments{
\item{data}{two-dimensional \code{matrix} of mode \code{numeric}; diversity
components.}

\item{qs}{\code{vector} of mode \code{numeric}; parameter of conservatism.}
}
\value{
Returns a five-column \code{tibble}/\code{dataframe} containing:
\code{partition} (label attributed to partition), \code{q} (parameter of
conservatism), \code{diversity}, \code{community} (level of diversity,
\emph{i.e.} subcommunity), and \code{measure} (alpha, beta, rho, or gamma).


}
\description{
Generic function for calculating subcommunity diversity.
}
\details{
\code{data} may be input as three different classes:
\itemize{
\item{\code{power_mean} calculates raw and normalised subcomunity alpha, rho
or gamma diversity by taking the powermean of diversity components}
\item{\code{relativeentropy} calculates raw or normalised subcommunity beta
diversity by taking the relative entropy of diversity components}
\item{\code{metacommunity} caculates all subcommunity measures of diversity}
}
}
\examples{
# Calculate the diversity of a single population
pop <- data.frame(a = c(1,3), b = c(1,1))
pop <- pop/sum(pop)
meta <- metacommunity(pop)

# Subcommunity gamma diversity (takes the power mean)
g <- raw_gamma(meta)
subdiv(g, 0:2)

# Subcommunity beta diversity (takes the relative entropy)
b <- raw_beta(meta)
subdiv(b, 0:2)

# All measures of subcommunity diversity
subdiv(meta, 0:2)

}

