% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-supercommunity.R, R/as.supercommunity.R
\docType{class}
\name{supercommunity-class}
\alias{is.supercommunity}
\alias{show,supercommunity-method}
\alias{supercommunity-class}
\title{supercommunity-class}
\usage{
is.supercommunity(x)

\S4method{show}{supercommunity}(object)
}
\arguments{
\item{x}{any R object}

\item{object}{object of class \code{supercommunity}}
}
\value{
returns TRUE if its argument is a supercommunity, FALSE otherwise.
}
\description{
Container for proportional abundance and similarity matrices.
}
\section{Fields}{

\describe{
\item{\code{.Data}}{two-dimensional \code{matrix} of mode \code{numeric}; contains 
proportional abundance of samples (usually types, except in the phylogenetic
case where samples correspond to the present day species)}

\item{\code{similarity}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains pairwise similarity between \emph{types}}

\item{\code{type_abundance}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains proportional abundance of \emph{types} in the subcommunity 
as a fraction of the supercommunity as a whole (in the phylogenetic case, 
this corresponds to the proportional abundance of historic species, which
is calculated from the proportional abundance of present day species)}

\item{\code{ordinariness}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains ordinariness of types}

\item{\code{subcommunity_weights}}{\code{vector} of mode \code{numeric}; contains
subcommunity weights}

\item{\code{type_weights}}{two-dimensional \code{matrix} of mode \code{numeric}; 
contains weight of types within a subcommunity}
}}

