% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.supercommunity.R
\docType{methods}
\name{supercommunity}
\alias{ANY-method}
\alias{as.supercommunity}
\alias{supercommunity}
\alias{supercommunity,}
\alias{supercommunity,data.frame,matrix-method}
\alias{supercommunity,data.frame,missing-method}
\alias{supercommunity,data.frame,phylo-method}
\alias{supercommunity,data.frame,rdphylo-method}
\alias{supercommunity,matrix,matrix-method}
\alias{supercommunity,matrix,missing-method}
\alias{supercommunity,matrix,phylo-method}
\alias{supercommunity,matrix,rdphylo-method}
\alias{supercommunity,missing,phylo-method}
\alias{supercommunity,missing,rdphylo-method}
\alias{supercommunity,numeric,matrix-method}
\alias{supercommunity,numeric,missing-method}
\alias{supercommunity,numeric,phylo-method}
\alias{supercommunity,numeric,rdphylo-method}
\title{Coerce to Supercommunity}
\usage{
supercommunity(partition, similarity, ...)

\S4method{supercommunity}{data.frame,missing}(partition, similarity = NA)

\S4method{supercommunity}{numeric,missing}(partition, similarity = NA)

\S4method{supercommunity}{data.frame,matrix}(partition, similarity = NA)

\S4method{supercommunity}{numeric,matrix}(partition, similarity = NA)

\S4method{supercommunity}{matrix,missing}(partition, similarity = NA)

\S4method{supercommunity}{matrix,matrix}(partition, similarity = NA)

\S4method{supercommunity}{missing,phylo}(partition, similarity, interval)

\S4method{supercommunity}{numeric,phylo}(partition, similarity, interval)

\S4method{supercommunity}{data.frame,phylo}(partition, similarity, interval)

\S4method{supercommunity}{matrix,phylo}(partition, similarity, interval)

\S4method{supercommunity}{missing,rdphylo}(partition, similarity, interval)

\S4method{supercommunity}{numeric,rdphylo}(partition, similarity, interval)

\S4method{supercommunity}{data.frame,rdphylo}(partition, similarity, interval)

\S4method{supercommunity}{matrix,rdphylo}(partition, similarity, interval)

supercommunity(partition, similarity, ...)
}
\arguments{
\item{partition}{(optional) \code{vector} of mode \code{numeric}; containing the 
proportional abundance of present day species (leaves)}

\item{similarity}{object of class \code{phylo} or \code{rdphylo}}

\item{...}{additional arguments.}

\item{interval}{object of mode numeric}
}
\value{
Returns an object of class \code{supercommunity}; an S4 object 
containing five slots (see Details).
}
\description{
Functions to check if an object is a \code{supercommunity} or coerce an  
object into a \code{supercommunity}.
}
\details{
\enumerate{
\item .Data (partition) - proportional abundance of samples (usually types, 
except in the phylogenetic case where samples correspond to the present day 
species)
\item similarity - pairwise similarity between \emph{types}
\item type_abundance - proportional abundance of \emph{types} in the  
subcommunity as a fraction of the supercommunity as a whole (in the  
phylogenetic case, this corresponds to the proportional abundance of 
historic species, which is calculated from the proportional abundance of 
present day species)
\item ordinariness - ordinariness of types
\item subcommunity_weights - subcommunity weights
\item type_weights - weight of types within a subcommunity
}
}
\examples{
tree <- ape::rtree(n = 5)
partition <- data.frame(a = c(1,0,1,0,0), b = c(0,1,0,1,1))
partition <- partition / sum(partition)
a <- supercommunity(partition, tree)
str(a)
a

showMethods("supercommunity")

}
\seealso{
\code{\link{supercommunity-class}}
}

