% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-measures.R
\name{supercommunity.A.bar}
\alias{supercommunity.A.bar}
\title{Similarity-sensitive normalised supercommunity.A diversity}
\usage{
supercommunity.A.bar(super, qs)
}
\arguments{
\item{super}{object of class \code{supercommunity}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
Returns a two-dimensional \code{matrix} of diversities, with columns  
representing subcommunities and rows representing values of \emph{q}.
}
\description{
\code{supercommunity.A.bar()} is used to calculate the 
average similarity-sensitive diversity of subcommunities.
}
\details{
\code{supercommunity.A.bar()} calculates the total normalised supercommunity  
alpha diversity of a series of columns representing independent subcommunity 
counts relative to the supercommunity as a whole (by default the sum of the 
subcommunities). This measure may be calculated for a series of orders, 
repesented as a vector of \code{qs}.
}
\examples{
pop <- sample(1:50, 5)
super <- supercommunity(pop)

# Calculate normalised supercommunity alpha diversity (takes the power mean)
a <- alphabar(super)
superdiv(a, 0:2)

}

