% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.A.bar.R
\name{supercommunity.A.bar}
\alias{supercommunity.A.bar}
\title{Similarity-sensitive Normalised supercommunity.A diversity}
\usage{
supercommunity.A.bar(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity alpha diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\seealso{
\code{\link{supercommunity.A}}, \code{\link{subcommunity.alpha}}, \code{\link{subcommunity.alpha.bar}}
}

