% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.A.bar.R
\name{supercommunity.A.bar}
\alias{supercommunity.A.bar}
\title{Similarity-sensitive Normalised supercommunity.A diversity}
\usage{
supercommunity.A.bar(populations, qs)
}
\arguments{
\item{populations}{object of class \code{collection}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity alpha diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Coerse object into a collection
data <- as.collection(population)

# Calculate diversity
supercommunity.A.bar(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

