% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.A.bar.R
\name{supercommunity.A.bar}
\alias{supercommunity.A.bar}
\title{Similarity-sensitive Normalised supercommunity.A diversity}
\usage{
supercommunity.A.bar(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}
}
\value{
An array of diversities, last representing values of q
}
\description{
Calculates the total supercommunity alpha diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of qs.
}
\seealso{
\code{\link{supercommunity.A}}
}

