% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.A.R
\name{supercommunity.A}
\alias{supercommunity.A}
\title{Similarity-sensitive Raw supercommunity.A diversity}
\usage{
supercommunity.A(populations, qs, Z = diag(nrow(populations)),
  normalise = F)
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}

\item{normalise}{Normalise probability distribution to sum to 1}
}
\value{
An array of diversities, last representing values of q
}
\description{
Calculates the total supercommunity alpha diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of qs.
}
\seealso{
\code{\link{supercommunity.A.bar}}
}

