% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.A.R
\name{supercommunity.A}
\alias{supercommunity.A}
\title{Similarity-sensitive Raw supercommunity.A diversity}
\usage{
supercommunity.A(populations, qs, normalise = F)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity alpha diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace=T),
                        subcommunityB = sample(1:50, 5, replace=T))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

# Calculate diversity
supercommunity.A(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

