% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.B.bar.R
\name{supercommunity.B.bar}
\alias{supercommunity.B.bar}
\title{Similarity-sensitive Normalised supercommunity.B diversity}
\usage{
supercommunity.B.bar(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity beta diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\seealso{
\code{\link{supercommunity.B}}, \code{\link{subcommunity.beta}}, \code{\link{subcommunity.beta.bar}}
}

