% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.B.bar.R
\name{supercommunity.B.bar}
\alias{supercommunity.B.bar}
\title{Similarity-sensitive Normalised supercommunity.B diversity}
\usage{
supercommunity.B.bar(populations, qs)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity beta diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace=T),
                        subcommunityB = sample(1:50, 5, replace=T))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

# Calculate diversity
supercommunity.B.bar(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

