% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.B.R
\name{supercommunity.B}
\alias{supercommunity.B}
\title{Similarity-sensitive Raw supercommunity.B diversity}
\usage{
supercommunity.B(populations, qs, normalise = F)
}
\arguments{
\item{populations}{object of class \code{collection}}

\item{qs}{\code{vector} of \emph{q} values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity beta diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\details{
\code{subcommunity.B()} is used to calculate the 
average distinctiveness of subcommunities.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Coerse object into a supercommunity
data <- as.supercommunity(population)

# Calculate diversity
supercommunity.B(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

