% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.G.bar.R
\name{supercommunity.G.bar}
\alias{supercommunity.G.bar}
\title{Similarity-sensitive Normalised supercommunity.G diversity}
\usage{
supercommunity.G.bar(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}
}
\value{
array of diversities, last representing values of q
}
\description{
Calculates the total supercommunity gamma diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of qs.
}
\seealso{
\code{\link{supercommunity.G}}
}

