% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.G.bar.R
\name{supercommunity.G.bar}
\alias{supercommunity.G.bar}
\title{Similarity-sensitive Normalised supercommunity.G diversity}
\usage{
supercommunity.G.bar(populations, qs)
}
\arguments{
\item{populations}{object of class \code{collection}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
array of diversities, last representing values of \emph{q}
}
\description{
\code{supercommunity.G()} is used to calculate the 
supercommunity similarity-sensitive diversity.
}
\details{
\code{supercommunity.G()} calculates the total supercommunity gamma diversity 
of a series of columns representing subcommunity counts, for a series of 
orders, repesented as a vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Coerse object into a supercommunity
data <- supercommunity(population)

# Calculate diversity
supercommunity.G.bar(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

