% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.G.R
\name{supercommunity.G}
\alias{supercommunity.G}
\title{Similarity-sensitive Raw supercommunity.G diversity}
\usage{
supercommunity.G(populations, qs, Z = diag(nrow(populations)),
  normalise = F)
}
\arguments{
\item{populations}{\eqn{n x m} \code{matrix} of population counts or proportions}

\item{qs}{\code{vector} of \emph{q} values}

\item{Z}{\eqn{n x n} \code{matrix} of pair-wise similarity values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity gamma diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\seealso{
\code{\link{supercommunity.G.bar}}, \code{\link{subcommunity.gamma}}, \code{\link{subcommunity.gamma.bar}}
}

