% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.G.R
\name{supercommunity.G}
\alias{supercommunity.G}
\title{Similarity-sensitive Raw supercommunity.G diversity}
\usage{
supercommunity.G(populations, qs, normalise = F)
}
\arguments{
\item{populations}{object of class \code{initDiv}}

\item{qs}{\code{vector} of \emph{q} values}

\item{normalise}{logical operator; T returns normalised probability 
distribution summed to 1}
}
\value{
array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity gamma diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Create object of class initDiv
data <- set.collection(population)

# Calculate diversity
supercommunity.G(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

