% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\docType{data}
\name{supercommunity.G}
\alias{supercommunity.G}
\title{Similarity-sensitive Raw supercommunity.G diversity}
\format{An object of class \code{diversity} of length 1.}
\usage{
supercommunity.G
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}

\item{normalise}{Normalise probability distribution to sum to 1}
}
\value{
array of diversities, last representing values of q
}
\description{
Calculates the total supercommunity gamma diversity of a series of columns
representing subcommunity counts, for a series of orders, repesented as a 
vector of qs.
}
\keyword{datasets}

