% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.R.bar.R
\name{supercommunity.R.bar}
\alias{supercommunity.R.bar}
\title{Similarity-sensitive Normalised supercommunity.R diversity}
\usage{
supercommunity.R.bar(populations, qs, Z = diag(nrow(populations)))
}
\arguments{
\item{populations}{Population counts or proportions}

\item{qs}{Vector of values of parameter q}

\item{Z}{Similarity matrix}
}
\value{
An array of diversities, last representing values of q
}
\description{
Calculates the total supercommunity.R.bar diversity of a series of 
columns representing subcommunity counts, for a series of orders, 
repesented as a vector of qs.
}
\seealso{
\code{\link{supercommunity.R}}
}

