% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercommunity.R.bar.R
\name{supercommunity.R.bar}
\alias{supercommunity.R.bar}
\title{Similarity-sensitive Normalised supercommunity.R diversity}
\usage{
supercommunity.R.bar(populations, qs)
}
\arguments{
\item{populations}{object of class \code{collection}}

\item{qs}{\code{vector} of \emph{q} values}
}
\value{
An array of diversities, last representing values of \emph{q}
}
\description{
Calculates the total supercommunity.R.bar diversity of a series of 
columns representing subcommunity counts, for a series of orders, 
repesented as a vector of \code{qs}.
}
\examples{
# Species counts
population <- data.frame(subcommunityA = sample(1:50, 5, replace = TRUE),
                        subcommunityB = sample(1:50, 5, replace = TRUE))
row.names(population) <- c('cows', 'sheep', 'ducks', 'foxes', 'bears')

# Coerse object into a supercommunity
data <- as.supercommunity(population)

# Calculate diversity
supercommunity.R.bar(data, 0:2)

}
\seealso{
\code{\link{diversity}}
}

