% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superdiv.R
\docType{methods}
\name{superdiv}
\alias{superdiv}
\alias{superdiv,powermean-method}
\alias{superdiv,relativeentropy-method}
\title{Calculate supercommunity diversity}
\usage{
superdiv(data, qs)

\S4method{superdiv}{powermean}(data, qs)

\S4method{superdiv}{relativeentropy}(data, qs)
}
\arguments{
\item{data}{two-dimensional \code{matrix} of mode \code{numeric}; diversity 
terms.}

\item{qs}{\code{vector} of mode \code{numeric}; parameter of conservatism.}
}
\value{
Returns a two-dimensional \code{matrix} of mode \code{numeric}.
}
\description{
Calculates supercommunity diversity by taking the powermean of diversity 
terms.
}
\details{

}
\examples{



}

