% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax2dist.R
\name{tax2dist}
\alias{tax2dist}
\title{Generate taxonomic distance matrix}
\usage{
tax2dist(lookup, values = c(species = 0, genus = 1, family = 2, subclass = 3,
  other = 4))
}
\arguments{
\item{lookup}{\code{data.frame} with colnames corresponding to nested 
hierarchical levels, e.g. c('Species', 'Genus', 'Family', 'Subclass')}

\item{values}{\code{vector} with of values of similarity attributed to 
hierarchical levels defined in \code{lookup}. Default is Shimatani's 
taxonomic distance parameters.}
}
\value{
\code{tax2dist()} returns a \code{matrix} of pair-wise taxonomic distances
}
\description{
Calculates taxonomic distances between species. By default these are based 
on Shimatani's taxonomic distance parameters (see \emph{Details}).
}
\details{
By default:
Individuals of the same species share a distance of 0 \cr
Individuals of the same genus but different species share a distance of 0.25 \cr
Individuals of the same family but different genus share a distance of 0.5 \cr
Individuals of the same subclass but different family share a distance of 0.75 \cr
Individuals of different subclass share a distance of 1
}
\examples{
# Create Lookup table
Species <- c("tenuifolium", "asterolepis", "simplex var.grandiflora", "simplex var.ochnacea")
Genus <- c("Protium", "Quararibea", "Swartzia", "Swartzia")
Family <- c("Burseraceae", "Bombacaceae", "Fabaceae", "Fabaceae")
Subclass <- c("Sapindales", "Malvales", "Fabales", "Fabales")
lookup <- cbind.data.frame(Species, Genus, Family, Subclass)

# Assign values for each level
values <- c(species = 0, genus = 1, family = 2, subclass = 3, other = 4)

# Generate pairwise distances
dist <- tax2dist(lookup, values)

# Convert distances to similarities
dist2sim(dist, "l")

}
\references{
Shimatani, K. 2001. On the measurement of species diversity 
incorporating species differences. Oikos 93:135–147.
}
