% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxfac.R
\name{taxfac}
\alias{taxfac}
\title{taxfac}
\usage{
taxfac(lookup)
}
\arguments{
\item{lookup}{\code{data.frame} with colnames corresponding to nested
hierarchical levels, e.g. c('Species', 'Genus', 'Family', 'Subclass')}
}
\description{
taxfac
}
\examples{
\dontrun{
# Create Lookup table
Species <- c("tenuifolium", "asterolepis", "simplex var.grandiflora", "simplex var.ochnacea")
Genus <- c("Protium", "Quararibea", "Swartzia", "Swartzia")
Family <- c("Burseraceae", "Bombacaceae", "Fabaceae", "Fabaceae")
Subclass <- c("Sapindales", "Malvales", "Fabales", "Fabales")
lookup <- cbind.data.frame(Species, Genus, Family, Subclass)

taxfac(lookup)
}

}
