% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxid.R
\name{taxid}
\alias{taxid}
\title{taxid}
\usage{
taxid(taxFac)
}
\arguments{
\item{taxFac}{sdf}
}
\description{
Generate taxanomic codes for each species by converting species, genus, 
family, and subclass into factors
}
