% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Cut phylogeny}
\usage{
trim(tree, partition, interval)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with 
rows as types, columns as subcommunities, and elements containing relative 
abundances of types in subcommunities. In the case of phylogenetic 
metacommunities, these are the relative abundances of terminal taxa.}

\item{interval}{proportion of total tree height to be conserved (taken as
a proportion from the heighest tip). Describes how far back we go in the tree,
with 0 marking the date of the most recent tip, and 1 (the default) marking 
the most recent common ancestor. Numbers greater than 1 extend the root of 
the tree.}
}
\value{
Returns an object of class \code{phy_struct} containing the original 
phylogenetic parameters ('@parameters') and a new structural matrix 
('@structure').
}
\description{
Cut phylogeny
}
\examples{
tree <- ape::rtree(n = 5)
tree$tip.label <- paste0("sp", seq_along(tree$tip.label))

partition <- cbind(a = c(0,1,1,1,0), b = c(1,0,0,0,1))
row.names(partition) <- tree$tip.label 
partition <- partition / sum(partition)

res <- trim(tree, partition, 0.2)

}
