% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrix.R
\name{z_matrix}
\alias{z_matrix}
\title{z_matrix}
\usage{
z_matrix(partition, smatrix, ps)
}
\arguments{
\item{partition}{two-dimensinal \code{matrix} of mode \code{numeric} with 
rows as types, columns as subcommunities, and elements containing relative 
abundances of types in subcommunities. In the case of phylogenetic 
metacommunities, these are the relative abundances of terminal taxa.}

\item{smatrix}{object}

\item{ps}{object}
}
\value{
\eqn{hS x hS} matrix; pair-wise similarity of historic species
}
\description{
Function to
}
\examples{
tree <- ape::read.tree(text="(A:1,B:2)R:1;")
partition <- setNames(c(0.6, 0.4), tree$tip.label)
ps <- phy_struct(partition, tree)
smatrix <- s_matrix(tree, ps)
z_matrix(partition, smatrix, ps)

}
